% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day18.R
\name{bfs3d}
\alias{bfs3d}
\title{adapted from https://favtutor.com/blogs/breadth-first-search-python}
\usage{
bfs3d(node, x, neighbours, maxnodes)
}
\description{
adapted from https://favtutor.com/blogs/breadth-first-search-python
}
