% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{day04}
\alias{day04}
\alias{f04a}
\alias{f04b}
\alias{example_data_04}
\title{Day 04: Scratchcards}
\usage{
f04a(x)

f04b(x)

example_data_04(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f04a(x)} returns .... For Part Two,
\code{f04b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2023/day/4}{Scratchcards}
}
\details{
\strong{Part One}

The gondola takes you up. Strangely, though, the ground doesn\'t seem to
be coming with you; you\'re not climbing a mountain. As the circle of
Snow Island recedes below you, an entire new landmass suddenly appears
above you! The gondola carries you to the surface of the new island and
lurches into the station.

As you exit the gondola, the first thing you notice is that the air here
is much \emph{warmer} than it was on Snow Island. It\'s also quite \emph{humid}.
Is this where the water source is?

The next thing you notice is an Elf sitting on the floor across the
station in what seems to be a pile of colorful square cards.

\"Oh! Hello!\" The Elf excitedly runs over to you. \"How may I be of
service?\" You ask about water sources.

\"I\'m not sure; I just operate the gondola lift. That does sound like
something we\'d have, though - this is \emph{[Island
Island]{title="I find this way funnier than it has any right to be."}},
after all! I bet the \emph{gardener} would know. He\'s on a different island,
though - er, the small kind surrounded by water, not the floating kind.
We really need to come up with a better naming scheme. Tell you what: if
you can help me with something quick, I\'ll let you \emph{borrow my boat} and
you can go visit the gardener. I got all these
\href{https://en.wikipedia.org/wiki/Scratchcard}{scratchcards}{target="_blank"}
as a gift, but I can\'t figure out what I\'ve won.\"

The Elf leads you over to the pile of colorful cards. There, you
discover dozens of scratchcards, all with their opaque covering already
scratched off. Picking one up, it looks like each card has two lists of
numbers separated by a vertical bar (\code{|}): a list of \emph{winning numbers}
and then a list of \emph{numbers you have}. You organize the information into
a table (your puzzle input).

As far as the Elf has been able to figure out, you have to figure out
which of the \emph{numbers you have} appear in the list of \emph{winning numbers}.
The first match makes the card worth \emph{one point} and each match after
the first \emph{doubles} the point value of that card.

For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Card 1: 41 48 83 86 17 | 83 86  6 31 17  9 48 53
Card 2: 13 32 20 16 61 | 61 30 68 82 17 32 24 19
Card 3:  1 21 53 59 44 | 69 82 63 72 16 21 14  1
Card 4: 41 92 73 84 69 | 59 84 76 51 58  5 54 83
Card 5: 87 83 26 28 32 | 88 30 70 12 93 22 82 36
Card 6: 31 18 13 56 72 | 74 77 10 23 35 67 36 11
}\if{html}{\out{</div>}}

In the above example, card 1 has five winning numbers (\code{41}, \code{48}, \code{83},
\code{86}, and \code{17}) and eight numbers you have (\code{83}, \code{86}, \code{6}, \code{31}, \code{17},
\code{9}, \code{48}, and \code{53}). Of the numbers you have, four of them (\code{48}, \code{83},
\code{17}, and \code{86}) are winning numbers! That means card 1 is worth \emph{\code{8}}
points (1 for the first match, then doubled three times for each of the
three matches after the first).
\itemize{
\item Card 2 has two winning numbers (\code{32} and \code{61}), so it is worth \emph{\code{2}}
points.
\item Card 3 has two winning numbers (\code{1} and \code{21}), so it is worth \emph{\code{2}}
points.
\item Card 4 has one winning number (\code{84}), so it is worth \emph{\code{1}} point.
\item Card 5 has no winning numbers, so it is worth no points.
\item Card 6 has no winning numbers, so it is worth no points.
}

So, in this example, the Elf\'s pile of scratchcards is worth \emph{\code{13}}
points.

Take a seat in the large pile of colorful cards. \emph{How many points are
they worth in total?}

\strong{Part Two}

Just as you\'re about to report your findings to the Elf, one of you
realizes that the rules have actually been printed on the back of every
card this whole time.

There\'s no such thing as \"points\". Instead, scratchcards only cause
you to \emph{win more scratchcards} equal to the number of winning numbers
you have.

Specifically, you win \emph{copies} of the scratchcards below the winning
card equal to the number of matches. So, if card 10 were to have 5
matching numbers, you would win one copy each of cards 11, 12, 13, 14,
and 15.

Copies of scratchcards are scored like normal scratchcards and have the
\emph{same card number} as the card they copied. So, if you win a copy of
card 10 and it has 5 matching numbers, it would then win a copy of the
same cards that the original card 10 won: cards 11, 12, 13, 14, and 15.
This process repeats until none of the copies cause you to win any more
cards. (Cards will never make you copy a card past the end of the
table.)

This time, the above example goes differently:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Card 1: 41 48 83 86 17 | 83 86  6 31 17  9 48 53
Card 2: 13 32 20 16 61 | 61 30 68 82 17 32 24 19
Card 3:  1 21 53 59 44 | 69 82 63 72 16 21 14  1
Card 4: 41 92 73 84 69 | 59 84 76 51 58  5 54 83
Card 5: 87 83 26 28 32 | 88 30 70 12 93 22 82 36
Card 6: 31 18 13 56 72 | 74 77 10 23 35 67 36 11
}\if{html}{\out{</div>}}
\itemize{
\item Card 1 has four matching numbers, so you win one copy each of the
next four cards: cards 2, 3, 4, and 5.
\item Your original card 2 has two matching numbers, so you win one copy
each of cards 3 and 4.
\item Your copy of card 2 also wins one copy each of cards 3 and 4.
\item Your four instances of card 3 (one original and three copies) have
two matching numbers, so you win \emph{four} copies each of cards 4 and
5.
\item Your eight instances of card 4 (one original and seven copies) have
one matching number, so you win \emph{eight} copies of card 5.
\item Your fourteen instances of card 5 (one original and thirteen copies)
have no matching numbers and win no more cards.
\item Your one instance of card 6 (one original) has no matching numbers
and wins no more cards.
}

Once all of the originals and copies have been processed, you end up
with \emph{\code{1}} instance of card 1, \emph{\code{2}} instances of card 2, \emph{\code{4}}
instances of card 3, \emph{\code{8}} instances of card 4, \emph{\code{14}} instances of card
5, and \emph{\code{1}} instance of card 6. In total, this example pile of
scratchcards causes you to ultimately have \emph{\code{30}} scratchcards!

Process all of the original and copied scratchcards until no more
scratchcards are won. Including the original set of scratchcards, \emph{how
many total scratchcards do you end up with?}
}
\examples{
f04a(example_data_04())
f04b()
}
