% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05}
\alias{day05}
\alias{f05a}
\alias{f05b}
\alias{example_data_05}
\title{Day 05: Supply Stacks}
\usage{
f05a(x, model = 9000)

f05b(x)

example_data_05(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f05a(x)} returns .... For Part Two,
\code{f05b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2022/day/5}{Supply Stacks}
}
\details{
\strong{Part One}

The expedition can depart as soon as the final supplies have been
unloaded from the ships. Supplies are stored in stacks of marked
\emph{crates}, but because the needed supplies are buried under many other
crates, the crates need to be rearranged.

The ship has a \emph{giant cargo crane} capable of moving crates between
stacks. To ensure none of the crates get crushed or fall over, the crane
operator will rearrange them in a series of carefully-planned steps.
After the crates are rearranged, the desired crates will be at the top
of each stack.

The Elves don\'t want to interrupt the crane operator during this
delicate procedure, but they forgot to ask her \emph{which} crate will end up
where, and they want to be ready to unload them as soon as possible so
they can embark.

They do, however, have a drawing of the starting stacks of crates \emph{and}
the rearrangement procedure (your puzzle input). For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    [D]
[N] [C]
[Z] [M] [P]
 1   2   3

move 1 from 2 to 1
move 3 from 1 to 3
move 2 from 2 to 1
move 1 from 1 to 2
}\if{html}{\out{</div>}}

In this example, there are three stacks of crates. Stack 1 contains two
crates: crate \code{Z} is on the bottom, and crate \code{N} is on top. Stack 2
contains three crates; from bottom to top, they are crates \code{M}, \code{C}, and
\code{D}. Finally, stack 3 contains a single crate, \code{P}.

Then, the rearrangement procedure is given. In each step of the
procedure, a quantity of crates is moved from one stack to a different
stack. In the first step of the above rearrangement procedure, one crate
is moved from stack 2 to stack 1, resulting in this configuration:

\if{html}{\out{<div class="sourceCode">}}\preformatted{[D]
[N] [C]
[Z] [M] [P]
 1   2   3
}\if{html}{\out{</div>}}

In the second step, three crates are moved from stack 1 to stack 3.
Crates are moved \emph{one at a time}, so the first crate to be moved (\code{D})
ends up below the second and third crates:

\if{html}{\out{<div class="sourceCode">}}\preformatted{        [Z]
        [N]
    [C] [D]
    [M] [P]
 1   2   3
}\if{html}{\out{</div>}}

Then, both crates are moved from stack 2 to stack 1. Again, because
crates are moved \emph{one at a time}, crate \code{C} ends up below crate \code{M}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{        [Z]
        [N]
[M]     [D]
[C]     [P]
 1   2   3
}\if{html}{\out{</div>}}

Finally, one crate is moved from stack 1 to stack 2:

\if{html}{\out{<div class="sourceCode">}}\preformatted{        [Z]
        [N]
        [D]
[C] [M] [P]
 1   2   3
}\if{html}{\out{</div>}}

The Elves just need to know \emph{which crate will end up on top of each
stack}; in this example, the top crates are \code{C} in stack 1, \code{M} in stack
2, and \code{Z} in stack 3, so you should combine these together and give the
Elves the message \emph{\code{CMZ}}.

\emph{After the rearrangement procedure completes, what crate ends up on top
of each stack?}

\strong{Part Two}

As you watch the crane operator expertly rearrange the crates, you
notice the process isn\'t following your prediction.

Some mud was covering the writing on the side of the crane, and you
quickly wipe it away. The crane isn\'t a CrateMover 9000 - it\'s a
\emph{[CrateMover
9001]{title="It's way better than the old CrateMover 1006."}}.

The CrateMover 9001 is notable for many new and exciting features: air
conditioning, leather seats, an extra cup holder, and \emph{the ability to
pick up and move multiple crates at once}.

Again considering the example above, the crates begin in the same
configuration:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    [D]
[N] [C]
[Z] [M] [P]
 1   2   3
}\if{html}{\out{</div>}}

Moving a single crate from stack 2 to stack 1 behaves the same as
before:

\if{html}{\out{<div class="sourceCode">}}\preformatted{[D]
[N] [C]
[Z] [M] [P]
 1   2   3
}\if{html}{\out{</div>}}

However, the action of moving three crates from stack 1 to stack 3 means
that those three moved crates \emph{stay in the same order}, resulting in
this new configuration:

\if{html}{\out{<div class="sourceCode">}}\preformatted{        [D]
        [N]
    [C] [Z]
    [M] [P]
 1   2   3
}\if{html}{\out{</div>}}

Next, as both crates are moved from stack 2 to stack 1, they \emph{retain
their order} as well:

\if{html}{\out{<div class="sourceCode">}}\preformatted{        [D]
        [N]
[C]     [Z]
[M]     [P]
 1   2   3
}\if{html}{\out{</div>}}

Finally, a single crate is still moved from stack 1 to stack 2, but now
it\'s crate \code{C} that gets moved:

\if{html}{\out{<div class="sourceCode">}}\preformatted{        [D]
        [N]
        [Z]
[M] [C] [P]
 1   2   3
}\if{html}{\out{</div>}}

In this example, the CrateMover 9001 has put the crates in a totally
different order: \emph{\code{MCD}}.

Before the rearrangement process finishes, update your simulation so
that the Elves know where they should stand to be ready to unload the
final supplies. \emph{After the rearrangement procedure completes, what crate
ends up on top of each stack?}
}
\examples{
f05a(example_data_05())
f05b()
}
