% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\name{day06}
\alias{day06}
\alias{f06a}
\alias{f06b}
\alias{example_data_06}
\title{Day 06: Wait For It}
\usage{
f06a(x)

f06b(x)

example_data_06(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f06a(x)} returns .... For Part Two,
\code{f06b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2023/day/6}{Wait For It}
}
\details{
\strong{Part One}

The ferry quickly brings you across Island Island. After asking around,
you discover that there is indeed normally a large pile of sand
somewhere near here, but you don\'t see anything besides lots of water
and the small island where the ferry has docked.

As you try to figure out what to do next, you notice a poster on a wall
near the ferry dock. \"Boat races! Open to the public! Grand prize is an
all-expenses-paid trip to \emph{Desert Island}!\" That must be where the sand
comes from! Best of all, the boat races are starting in just a few
minutes.

You manage to sign up as a competitor in the boat races just in time.
The organizer explains that it\'s not really a traditional race -
instead, you will get a fixed amount of time during which your boat has
to travel as far as it can, and you win if your boat goes the farthest.

As part of signing up, you get a sheet of paper (your puzzle input) that
lists the \emph{time} allowed for each race and also the best \emph{distance} ever
recorded in that race. To guarantee you win the grand prize, you need to
make sure you \emph{go farther in each race} than the current record holder.

The organizer brings you over to the area where the boat races are held.
The boats are much smaller than you expected - they\'re actually \emph{toy
boats}, each with a big button on top. Holding down the button \emph{charges
the boat}, and releasing the button \emph{allows the boat to move}. Boats
move faster if their button was held longer, but time spent holding the
button counts against the total race time. You can only hold the button
at the start of the race, and boats don\'t move until the button is
released.

For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Time:      7  15   30
Distance:  9  40  200
}\if{html}{\out{</div>}}

This document describes three races:
\itemize{
\item The first race lasts 7 milliseconds. The record distance in this
race is 9 millimeters.
\item The second race lasts 15 milliseconds. The record distance in this
race is 40 millimeters.
\item The third race lasts 30 milliseconds. The record distance in this
race is 200 millimeters.
}

Your toy boat has a starting speed of \emph{zero millimeters per
millisecond}. For each whole millisecond you spend at the beginning of
the race holding down the button, the boat\'s speed increases by \emph{one
millimeter per millisecond}.

So, because the first race lasts 7 milliseconds, you only have a few
options:
\itemize{
\item Don\'t hold the button at all (that is, hold it for \emph{\code{0}
milliseconds}) at the start of the race. The boat won\'t move; it
will have traveled \emph{\code{0} millimeters} by the end of the race.
\item Hold the button for \emph{\code{1} millisecond} at the start of the race.
Then, the boat will travel at a speed of \code{1} millimeter per
millisecond for 6 milliseconds, reaching a total distance traveled
of \emph{\code{6} millimeters}.
\item Hold the button for \emph{\code{2} milliseconds}, giving the boat a speed of
\code{2} millimeters per millisecond. It will then get 5 milliseconds to
move, reaching a total distance of \emph{\code{10} millimeters}.
\item Hold the button for \emph{\code{3} milliseconds}. After its remaining 4
milliseconds of travel time, the boat will have gone \emph{\code{12}
millimeters}.
\item Hold the button for \emph{\code{4} milliseconds}. After its remaining 3
milliseconds of travel time, the boat will have gone \emph{\code{12}
millimeters}.
\item Hold the button for \emph{\code{5} milliseconds}, causing the boat to travel a
total of \emph{\code{10} millimeters}.
\item Hold the button for \emph{\code{6} milliseconds}, causing the boat to travel a
total of \emph{\code{6} millimeters}.
\item Hold the button for \emph{\code{7} milliseconds}. That\'s the entire duration
of the race. You never let go of the button. The boat can\'t move
until you let you of the button. Please make sure you let go of the
button so the boat gets to move. \emph{\code{0} millimeters}.
}

Since the current record for this race is \code{9} millimeters, there are
actually \emph{\code{4}} different ways you could win: you could hold the button
for \code{2}, \code{3}, \code{4}, or \code{5} milliseconds at the start of the race.

In the second race, you could hold the button for at least \code{4}
milliseconds and at most \code{11} milliseconds and beat the record, a total
of \emph{\code{8}} different ways to win.

In the third race, you could hold the button for at least \code{11}
milliseconds and no more than \code{19} milliseconds and still beat the
record, a total of \emph{\code{9}} ways you could win.

To see how much margin of error you have, determine the \emph{number of ways
you can beat the record} in each race; in this example, if you multiply
these values together, you get \emph{\code{288}} (\code{4} \* \code{8} \* \code{9}).

Determine the number of ways you could beat the record in each race.
\emph{What do you get if you multiply these numbers together?}

\strong{Part Two}

As the race is about to start, you realize the piece of paper with race
times and record distances you got earlier actually just has [very
bad]{title="Keming!"}
\href{https://en.wikipedia.org/wiki/Kerning}{kerning}{target="_blank"}.
There\'s really \emph{only one race} - ignore the spaces between the numbers
on each line.

So, the example from before:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Time:      7  15   30
Distance:  9  40  200
}\if{html}{\out{</div>}}

\...now instead means this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Time:      71530
Distance:  940200
}\if{html}{\out{</div>}}

Now, you have to figure out how many ways there are to win this single
race. In this example, the race lasts for \emph{\code{71530} milliseconds} and the
record distance you need to beat is \emph{\code{940200} millimeters}. You could
hold the button anywhere from \code{14} to \code{71516} milliseconds and beat the
record, a total of \emph{\code{71503}} ways!

\emph{How many ways can you beat the record in this one much longer race?}
}
\examples{
f06a(example_data_06())
f06b()
}
