% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day08.R
\name{day08}
\alias{day08}
\alias{f08a}
\alias{f08b}
\alias{example_data_08}
\title{Day 08: Treetop Tree House}
\usage{
f08a(x)

f08b(x)

example_data_08(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f08a(x)} returns .... For Part Two,
\code{f08b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2022/day/8}{Treetop Tree House}
}
\details{
\strong{Part One}

The expedition comes across a peculiar patch of tall trees all planted
carefully in a grid. The Elves explain that a previous expedition
planted these trees as a reforestation effort. Now, they\'re curious if
this would be a good location for a \href{https://en.wikipedia.org/wiki/Tree_house}{tree house}{target="_blank"}.

First, determine whether there is enough tree cover here to keep a tree
house \emph{hidden}. To do this, you need to count the number of trees that
are \emph{visible from outside the grid} when looking directly along a row or
column.

The Elves have already launched a
\href{https://en.wikipedia.org/wiki/Quadcopter}{quadcopter}{target="_blank"}
to generate a map with the height of each tree ([your puzzle
input]{title="The Elves have already launched a quadcopter (your puzzle input)."}).
For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{30373
25512
65332
33549
35390
}\if{html}{\out{</div>}}

Each tree is represented as a single digit whose value is its height,
where \code{0} is the shortest and \code{9} is the tallest.

A tree is \emph{visible} if all of the other trees between it and an edge of
the grid are \emph{shorter} than it. Only consider trees in the same row or
column; that is, only look up, down, left, or right from any given tree.

All of the trees around the edge of the grid are \emph{visible} - since they
are already on the edge, there are no trees to block the view. In this
example, that only leaves the \emph{interior nine trees} to consider:
\itemize{
\item The top-left \code{5} is \emph{visible} from the left and top. (It isn\'t
visible from the right or bottom since other trees of height \code{5} are
in the way.)
\item The top-middle \code{5} is \emph{visible} from the top and right.
\item The top-right \code{1} is not visible from any direction; for it to be
visible, there would need to only be trees of height \emph{0} between it
and an edge.
\item The left-middle \code{5} is \emph{visible}, but only from the right.
\item The center \code{3} is not visible from any direction; for it to be
visible, there would need to be only trees of at most height \code{2}
between it and an edge.
\item The right-middle \code{3} is \emph{visible} from the right.
\item In the bottom row, the middle \code{5} is \emph{visible}, but the \code{3} and \code{4}
are not.
}

With 16 trees visible on the edge and another 5 visible in the interior,
a total of \emph{\code{21}} trees are visible in this arrangement.

Consider your map; \emph{how many trees are visible from outside the grid?}

\strong{Part Two}

Content with the amount of tree cover available, the Elves just need to
know the best spot to build their tree house: they would like to be able
to see a lot of \emph{trees}.

To measure the viewing distance from a given tree, look up, down, left,
and right from that tree; stop if you reach an edge or at the first tree
that is the same height or taller than the tree under consideration. (If
a tree is right on the edge, at least one of its viewing distances will
be zero.)

The Elves don\'t care about distant trees taller than those found by the
rules above; the proposed tree house has large
\href{https://en.wikipedia.org/wiki/Eaves}{eaves}{target="_blank"} to keep it
dry, so they wouldn\'t be able to see higher than the tree house anyway.

In the example above, consider the middle \code{5} in the second row:

\if{html}{\out{<div class="sourceCode">}}\preformatted{30373
25512
65332
33549
35390
}\if{html}{\out{</div>}}
\itemize{
\item Looking up, its view is not blocked; it can see \emph{\code{1}} tree (of
height \code{3}).
\item Looking left, its view is blocked immediately; it can see only \emph{\code{1}}
tree (of height \code{5}, right next to it).
\item Looking right, its view is not blocked; it can see \emph{\code{2}} trees.
\item Looking down, its view is blocked eventually; it can see \emph{\code{2}} trees
(one of height \code{3}, then the tree of height \code{5} that blocks its
view).
}

A tree\'s \emph{scenic score} is found by \emph{multiplying together} its viewing
distance in each of the four directions. For this tree, this is \emph{\code{4}}
(found by multiplying \code{1 * 1 * 2 * 2}).

However, you can do even better: consider the tree of height \code{5} in the
middle of the fourth row:

\if{html}{\out{<div class="sourceCode">}}\preformatted{30373
25512
65332
33549
35390
}\if{html}{\out{</div>}}
\itemize{
\item Looking up, its view is blocked at \emph{\code{2}} trees (by another tree with
a height of \code{5}).
\item Looking left, its view is not blocked; it can see \emph{\code{2}} trees.
\item Looking down, its view is also not blocked; it can see \emph{\code{1}} tree.
\item Looking right, its view is blocked at \emph{\code{2}} trees (by a massive tree
of height \code{9}).
}

This tree\'s scenic score is \emph{\code{8}} (\code{2 * 2 * 1 * 2}); this is the ideal
spot for the tree house.

Consider each tree on your map. \emph{What is the highest scenic score
possible for any tree?}
}
\examples{
f08a(example_data_08())
f08b()
}
