% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day09.R
\name{day09}
\alias{day09}
\alias{f09a}
\alias{f09b}
\alias{example_data_09}
\title{Day 09: Mirage Maintenance}
\usage{
f09a(x)

f09b(x)

example_data_09(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f09a(x)} returns .... For Part Two,
\code{f09b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2023/day/9}{Mirage Maintenance}
}
\details{
\strong{Part One}

You ride the camel through the sandstorm and stop where the ghost\'s
maps told you to stop. [The sandstorm subsequently subsides, somehow
seeing you standing at an
\emph{oasis}!]{title="The sound of a sandstorm slowly settling."}

The camel goes to get some water and you stretch your neck. As you look
up, you discover what must be yet another giant floating island, this
one made of metal! That must be where the \emph{parts to fix the sand
machines} come from.

There\'s even a \href{https://en.wikipedia.org/wiki/Hang_gliding}{hang glider}{target="_blank"}
partially buried in the sand here; once the sun rises and heats up the
sand, you might be able to use the glider and the hot air to get all the
way up to the metal island!

While you wait for the sun to rise, you admire the oasis hidden here in
the middle of Desert Island. It must have a delicate ecosystem; you
might as well take some ecological readings while you wait. Maybe you
can report any environmental instabilities you find to someone so the
oasis can be around for the next sandstorm-worn traveler.

You pull out your handy \emph{Oasis And Sand Instability Sensor} and analyze
your surroundings. The OASIS produces a report of many values and how
they are changing over time (your puzzle input). Each line in the report
contains the \emph{history} of a single value. For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{0 3 6 9 12 15
1 3 6 10 15 21
10 13 16 21 30 45
}\if{html}{\out{</div>}}

To best protect the oasis, your environmental report should include a
\emph{prediction of the next value} in each history. To do this, start by
making a new sequence from the \emph{difference at each step} of your
history. If that sequence is \emph{not} all zeroes, repeat this process,
using the sequence you just generated as the input sequence. Once all of
the values in your latest sequence are zeroes, you can extrapolate what
the next value of the original history should be.

In the above dataset, the first history is \verb{0 3 6 9 12 15}. Because the
values increase by \code{3} each step, the first sequence of differences that
you generate will be \verb{3 3 3 3 3}. Note that this sequence has one fewer
value than the input sequence because at each step it considers two
numbers from the input. Since these values aren\'t \emph{all zero}, repeat
the process: the values differ by \code{0} at each step, so the next sequence
is \verb{0 0 0 0}. This means you have enough information to extrapolate the
history! Visually, these sequences can be arranged like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{0   3   6   9  12  15
  3   3   3   3   3
    0   0   0   0
}\if{html}{\out{</div>}}

To extrapolate, start by adding a new zero to the end of your list of
zeroes; because the zeroes represent differences between the two values
above them, this also means there is now a placeholder in every sequence
above it:

\if{html}{\out{<div class="sourceCode">}}\preformatted{0   3   6   9  12  15   B
  3   3   3   3   3   A
    0   0   0   0   0
}\if{html}{\out{</div>}}

You can then start filling in placeholders from the bottom up. \code{A} needs
to be the result of increasing \code{3} (the value to its left) by \code{0} (the
value below it); this means \code{A} must be \emph{\code{3}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{0   3   6   9  12  15   B
  3   3   3   3   3   3
    0   0   0   0   0
}\if{html}{\out{</div>}}

Finally, you can fill in \code{B}, which needs to be the result of increasing
\code{15} (the value to its left) by \code{3} (the value below it), or \emph{\code{18}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{0   3   6   9  12  15  18
  3   3   3   3   3   3
    0   0   0   0   0
}\if{html}{\out{</div>}}

So, the next value of the first history is \emph{\code{18}}.

Finding all-zero differences for the second history requires an
additional sequence:

\if{html}{\out{<div class="sourceCode">}}\preformatted{1   3   6  10  15  21
  2   3   4   5   6
    1   1   1   1
      0   0   0
}\if{html}{\out{</div>}}

Then, following the same process as before, work out the next value in
each sequence from the bottom up:

\if{html}{\out{<div class="sourceCode">}}\preformatted{1   3   6  10  15  21  28
  2   3   4   5   6   7
    1   1   1   1   1
      0   0   0   0
}\if{html}{\out{</div>}}

So, the next value of the second history is \emph{\code{28}}.

The third history requires even more sequences, but its next value can
be found the same way:

\if{html}{\out{<div class="sourceCode">}}\preformatted{10  13  16  21  30  45  68
   3   3   5   9  15  23
     0   2   4   6   8
       2   2   2   2
         0   0   0
}\if{html}{\out{</div>}}

So, the next value of the third history is \emph{\code{68}}.

If you find the next value for each history in this example and add them
together, you get \emph{\code{114}}.

Analyze your OASIS report and extrapolate the next value for each
history. \emph{What is the sum of these extrapolated values?}

\strong{Part Two}

Of course, it would be nice to have \emph{even more history} included in your
report. Surely it\'s safe to just \emph{extrapolate backwards} as well,
right?

For each history, repeat the process of finding differences until the
sequence of differences is entirely zero. Then, rather than adding a
zero to the end and filling in the next values of each previous
sequence, you should instead add a zero to the \emph{beginning} of your
sequence of zeroes, then fill in new \emph{first} values for each previous
sequence.

In particular, here is what the third example history looks like when
extrapolating back in time:

\if{html}{\out{<div class="sourceCode">}}\preformatted{5  10  13  16  21  30  45
  5   3   3   5   9  15
   -2   0   2   4   6
      2   2   2   2
        0   0   0
}\if{html}{\out{</div>}}

Adding the new values on the left side of each sequence from bottom to
top eventually reveals the new left-most history value: \emph{\code{5}}.

Doing this for the remaining example data above results in previous
values of \emph{\code{-3}} for the first history and \emph{\code{0}} for the second history.
Adding all three new values together produces \emph{\code{2}}.

Analyze your OASIS report again, this time extrapolating the \emph{previous}
value for each history. \emph{What is the sum of these extrapolated values?}
}
\examples{
f09a(example_data_09())
f09b()
}
