% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day11.R
\name{day11}
\alias{day11}
\alias{f11a}
\alias{f11b}
\alias{example_data_11}
\title{Day 11: Cosmic Expansion}
\usage{
f11a(x)

f11b(x)

example_data_11(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f11a(x)} returns .... For Part Two,
\code{f11b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2023/day/11}{Cosmic Expansion}
}
\details{
\strong{Part One}

You continue following signs for \"Hot Springs\" and eventually come
across an
\href{https://en.wikipedia.org/wiki/Observatory}{observatory}{target="_blank"}.
The Elf within turns out to be a researcher studying cosmic expansion
using the giant telescope here.

He doesn\'t know anything about the missing machine parts; he\'s only
visiting for this research project. However, he confirms that the hot
springs are the next-closest area likely to have people; he\'ll even
take you straight there once he\'s done with today\'s observation
analysis.

Maybe you can help him with the analysis to speed things up?

The researcher has collected a bunch of data and compiled the data into
a single giant \emph{image} (your puzzle input). The image includes \emph{empty
space} (\code{.}) and \emph{galaxies} (\verb{#}). For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{...#......
.......#..
#.........
..........
......#...
.#........
.........#
..........
.......#..
#...#.....
}\if{html}{\out{</div>}}

The researcher is trying to figure out the sum of the lengths of the
\emph{shortest path between every pair of galaxies}. However, there\'s a
catch: the universe expanded in the time it took the light from those
galaxies to reach the observatory.

Due to something involving gravitational effects, \emph{only some space
expands}. In fact, the result is that \emph{any rows or columns that contain
no galaxies} should all actually be twice as big.

In the above example, three columns and two rows contain no galaxies:

\if{html}{\out{<div class="sourceCode">}}\preformatted{   v  v  v
 ...#......
 .......#..
 #.........
>..........<
 ......#...
 .#........
 .........#
>..........<
 .......#..
 #...#.....
   ^  ^  ^
}\if{html}{\out{</div>}}

These rows and columns need to be \emph{twice as big}; the result of cosmic
expansion therefore looks like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{....#........
.........#...
#............
.............
.............
........#....
.#...........
............#
.............
.............
.........#...
#....#.......
}\if{html}{\out{</div>}}

Equipped with this expanded universe, the shortest path between every
pair of galaxies can be found. It can help to assign every galaxy a
unique number:

\if{html}{\out{<div class="sourceCode">}}\preformatted{....1........
.........2...
3............
.............
.............
........4....
.5...........
............6
.............
.............
.........7...
8....9.......
}\if{html}{\out{</div>}}

In these 9 galaxies, there are \emph{36 pairs}. Only count each pair once;
order within the pair doesn\'t matter. For each pair, find any shortest
path between the two galaxies using only steps that move up, down, left,
or right exactly one \code{.} or \verb{#} at a time. (The shortest path between
two galaxies is allowed to pass through another galaxy.)

For example, here is one of the shortest paths between galaxies \code{5} and
\code{9}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{....1........
.........2...
3............
.............
.............
........4....
.5...........
.##.........6
..##.........
...##........
....##...7...
8....9.......
}\if{html}{\out{</div>}}

This path has length \emph{\code{9}} because it takes a minimum of \emph{nine steps} to
get from galaxy \code{5} to galaxy \code{9} (the eight locations marked \verb{#} plus
the step onto galaxy \code{9} itself). Here are some other example shortest
path lengths:
\itemize{
\item Between galaxy \code{1} and galaxy \code{7}: 15
\item Between galaxy \code{3} and galaxy \code{6}: 17
\item Between galaxy \code{8} and galaxy \code{9}: 5
}

In this example, after expanding the universe, the sum of the shortest
path between all 36 pairs of galaxies is \emph{\code{374}}.

Expand the universe, then find the length of the shortest path between
every pair of galaxies. \emph{What is the sum of these lengths?}

\strong{Part Two}

The galaxies are much \emph{older} (and thus much \emph{farther apart}) than the
researcher initially estimated.

Now, instead of the expansion you did before, make each empty row or
column \emph{[one
million]{title="And you have to have your pinky near your mouth when you do it."}
times} larger. That is, each empty row should be replaced with \code{1000000}
empty rows, and each empty column should be replaced with \code{1000000}
empty columns.

(In the example above, if each empty row or column were merely \code{10}
times larger, the sum of the shortest paths between every pair of
galaxies would be \emph{\code{1030}}. If each empty row or column were merely
\code{100} times larger, the sum of the shortest paths between every pair of
galaxies would be \emph{\code{8410}}. However, your universe will need to expand
far beyond these values.)

Starting with the same initial image, expand the universe according to
these new rules, then find the length of the shortest path between every
pair of galaxies. \emph{What is the sum of these lengths?}
}
\examples{
f11a(example_data_11())
f11b()
}
