% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day14.R
\name{day14}
\alias{day14}
\alias{f14a}
\alias{f14b}
\alias{example_data_14}
\title{Day 14: Regolith Reservoir}
\usage{
f14a(x)

f14b(x)

example_data_14(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f14a(x)} returns .... For Part Two,
\code{f14b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2022/day/14}{Regolith Reservoir}
}
\details{
\strong{Part One}

The distress signal leads you to a giant waterfall! Actually, hang on -
the signal seems like it\'s coming from the waterfall itself, and that
doesn\'t make any sense. However, you do notice a little path that leads
\emph{behind} the waterfall.

Correction: the distress signal leads you behind a giant waterfall!
There seems to be a large cave system here, and the signal definitely
leads further inside.

As you begin to make your way deeper underground, you feel the ground
rumble for a moment. Sand begins pouring into the cave! If you don\'t
quickly figure out where the sand is going, you could quickly become
trapped!

Fortunately, your \href{/2018/day/17}{familiarity} with analyzing the path of
falling material will come in handy here. You scan a two-dimensional
vertical slice of the cave above you (your puzzle input) and discover
that it is mostly \emph{air} with structures made of \emph{rock}.

Your scan traces the path of each solid rock structure and reports the
\verb{x,y} coordinates that form the shape of the path, where \code{x} represents
distance to the right and \code{y} represents distance down. Each path
appears as a single line of text in your scan. After the first point of
each path, each point indicates the end of a straight horizontal or
vertical line to be drawn from the previous point. For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{498,4 -> 498,6 -> 496,6
503,4 -> 502,4 -> 502,9 -> 494,9
}\if{html}{\out{</div>}}

This scan means that there are two paths of rock; the first path
consists of two straight lines, and the second path consists of three
straight lines. (Specifically, the first path consists of a line of rock
from \verb{498,4} through \verb{498,6} and another line of rock from \verb{498,6}
through \verb{496,6}.)

The sand is pouring into the cave from point \verb{500,0}.

Drawing rock as \verb{#}, air as \code{.}, and the source of the sand as \code{+}, this
becomes:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  4     5  5
  9     0  0
  4     0  3
0 ......+...
1 ..........
2 ..........
3 ..........
4 ....#...##
5 ....#...#.
6 ..###...#.
7 ........#.
8 ........#.
9 #########.
}\if{html}{\out{</div>}}

Sand is produced \emph{one unit at a time}, and the next unit of sand is not
produced until the previous unit of sand \emph{comes to rest}. A unit of sand
is large enough to fill one tile of air in your scan.

A unit of sand always falls \emph{down one step} if possible. If the tile
immediately below is blocked (by rock or sand), the unit of sand
attempts to instead move diagonally \emph{one step down and to the left}. If
that tile is blocked, the unit of sand attempts to instead move
diagonally \emph{one step down and to the right}. Sand keeps moving as long
as it is able to do so, at each step trying to move down, then
down-left, then down-right. If all three possible destinations are
blocked, the unit of sand \emph{comes to rest} and no longer moves, at which
point the next unit of sand is created back at the source.

So, drawing sand that has come to rest as \code{o}, the first unit of sand
simply falls straight down and then stops:

\if{html}{\out{<div class="sourceCode">}}\preformatted{......+...
..........
..........
..........
....#...##
....#...#.
..###...#.
........#.
......o.#.
#########.
}\if{html}{\out{</div>}}

The second unit of sand then falls straight down, lands on the first
one, and then comes to rest to its left:

\if{html}{\out{<div class="sourceCode">}}\preformatted{......+...
..........
..........
..........
....#...##
....#...#.
..###...#.
........#.
.....oo.#.
#########.
}\if{html}{\out{</div>}}

After a total of five units of sand have come to rest, they form this
pattern:

\if{html}{\out{<div class="sourceCode">}}\preformatted{......+...
..........
..........
..........
....#...##
....#...#.
..###...#.
......o.#.
....oooo#.
#########.
}\if{html}{\out{</div>}}

After a total of 22 units of sand:

\if{html}{\out{<div class="sourceCode">}}\preformatted{......+...
..........
......o...
.....ooo..
....#ooo##
....#ooo#.
..###ooo#.
....oooo#.
...ooooo#.
#########.
}\if{html}{\out{</div>}}

Finally, only two more units of sand can possibly come to rest:

\if{html}{\out{<div class="sourceCode">}}\preformatted{......+...
..........
......o...
.....ooo..
....#ooo##
...o#ooo#.
..###ooo#.
....oooo#.
.o.ooooo#.
#########.
}\if{html}{\out{</div>}}

Once all \emph{\code{24}} units of sand shown above have come to rest, all further
sand flows out the bottom, falling into the endless void. Just for fun,
the path any new sand takes before falling forever is shown here with
\code{~}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{.......+...
.......~...
......~o...
.....~ooo..
....~#ooo##
...~o#ooo#.
..~###ooo#.
..~..oooo#.
.~o.ooooo#.
~#########.
~..........
~..........
~..........
}\if{html}{\out{</div>}}

Using your scan, simulate the falling sand. \emph{How many units of sand come
to rest before sand starts flowing into the abyss below?}

\strong{Part Two}
You realize you misread the scan. There isn\'t an [endless
void]{title="Endless Void is my C cover band."} at the bottom of the
scan - there\'s floor, and you\'re standing on it!

You don\'t have time to scan the floor, so assume the floor is an
infinite horizontal line with a \code{y} coordinate equal to \emph{two plus the
highest \code{y} coordinate} of any point in your scan.

In the example above, the highest \code{y} coordinate of any point is \code{9},
and so the floor is at \code{y=11}. (This is as if your scan contained one
extra rock path like \verb{-infinity,11 -> infinity,11}.) With the added
floor, the example above now looks like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{        ...........+........
        ....................
        ....................
        ....................
        .........#...##.....
        .........#...#......
        .......###...#......
        .............#......
        .............#......
        .....#########......
        ....................
<-- etc #################### etc -->
}\if{html}{\out{</div>}}

To find somewhere safe to stand, you\'ll need to simulate falling sand
until a unit of sand comes to rest at \verb{500,0}, blocking the source
entirely and stopping the flow of sand into the cave. In the example
above, the situation finally looks like this after \emph{\code{93}} units of sand
come to rest:

\if{html}{\out{<div class="sourceCode">}}\preformatted{............o............
...........ooo...........
..........ooooo..........
.........ooooooo.........
........oo#ooo##o........
.......ooo#ooo#ooo.......
......oo###ooo#oooo......
.....oooo.oooo#ooooo.....
....oooooooooo#oooooo....
...ooo#########ooooooo...
..ooooo.......ooooooooo..
#########################
}\if{html}{\out{</div>}}

Using your scan, simulate the falling sand until the source of the sand
becomes blocked. \emph{How many units of sand come to rest?}
}
\examples{
f14a(example_data_14())
f14b()
}
