% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day17.R
\name{day17}
\alias{day17}
\alias{f17a}
\alias{f17b}
\alias{example_data_17}
\title{Day 17: Pyroclastic Flow}
\usage{
f17a(x, n, ret = c("floor", "heights", "chamber"))

f17b(x)

example_data_17(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f17a(x)} returns .... For Part Two,
\code{f17b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2022/day/17}{Pyroclastic Flow}
}
\details{
\strong{Part One}

Your handheld device has located an alternative exit from the cave for
you and the elephants. The ground is rumbling almost continuously now,
but the strange valves bought you some time. It\'s definitely getting
warmer in here, though.

The tunnels eventually open into a very tall, narrow chamber. Large,
oddly-shaped rocks are falling into the chamber from above, presumably
due to all the rumbling. If you can\'t work out where the rocks will
fall next, you might be
[crushed]{title="I am the man who arranges the blocks / that descend upon me from up above!"}!

The five types of rocks have the following peculiar shapes, where \verb{#} is
rock and \code{.} is empty space:

\if{html}{\out{<div class="sourceCode">}}\preformatted{####

.#.
###
.#.

..#
..#
###

#
#
#
#

##
##
}\if{html}{\out{</div>}}

The rocks fall in the order shown above: first the \code{-} shape, then the
\code{+} shape, and so on. Once the end of the list is reached, the same
order repeats: the \code{-} shape falls first, sixth, 11th, 16th, etc.

The rocks don\'t spin, but they do get pushed around by jets of hot gas
coming out of the walls themselves. A quick scan reveals the effect the
jets of hot gas will have on the rocks as they fall (your puzzle input).

For example, suppose this was the jet pattern in your cave:

\if{html}{\out{<div class="sourceCode">}}\preformatted{>>><<><>><<<>><>>><<<>>><<<><<<>><>><<>>
}\if{html}{\out{</div>}}

In jet patterns, \code{<} means a push to the left, while \code{>} means a push to
the right. The pattern above means that the jets will push a falling
rock right, then right, then right, then left, then left, then right,
and so on. If the end of the list is reached, it repeats.

The tall, vertical chamber is exactly \emph{seven units wide}. Each rock
appears so that its left edge is two units away from the left wall and
its bottom edge is three units above the highest rock in the room (or
the floor, if there isn\'t one).

After a rock appears, it alternates between \emph{being pushed by a jet of
hot gas} one unit (in the direction indicated by the next symbol in the
jet pattern) and then \emph{falling one unit down}. If any movement would
cause any part of the rock to move into the walls, floor, or a stopped
rock, the movement instead does not occur. If a \emph{downward} movement
would have caused a falling rock to move into the floor or an
already-fallen rock, the falling rock stops where it is (having landed
on something) and a new rock immediately begins falling.

Drawing falling rocks with \code{@} and stopped rocks with \verb{#}, the jet
pattern in the example above manifests as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{The first rock begins falling:
|..@@.|
|.......|
|.......|
|.......|
+-------+

Jet of gas pushes rock right:
|...@@|
|.......|
|.......|
|.......|
+-------+

Rock falls 1 unit:
|...@@|
|.......|
|.......|
+-------+

Jet of gas pushes rock right, but nothing happens:
|...@@|
|.......|
|.......|
+-------+

Rock falls 1 unit:
|...@@|
|.......|
+-------+

Jet of gas pushes rock right, but nothing happens:
|...@@|
|.......|
+-------+

Rock falls 1 unit:
|...@@|
+-------+

Jet of gas pushes rock left:
|..@@.|
+-------+

Rock falls 1 unit, causing it to come to rest:
|..####.|
+-------+

A new rock begins falling:
|...@...|
|..@@..|
|...@...|
|.......|
|.......|
|.......|
|..####.|
+-------+

Jet of gas pushes rock left:
|..@....|
|.@@...|
|..@....|
|.......|
|.......|
|.......|
|..####.|
+-------+

Rock falls 1 unit:
|..@....|
|.@@...|
|..@....|
|.......|
|.......|
|..####.|
+-------+

Jet of gas pushes rock right:
|...@...|
|..@@..|
|...@...|
|.......|
|.......|
|..####.|
+-------+

Rock falls 1 unit:
|...@...|
|..@@..|
|...@...|
|.......|
|..####.|
+-------+

Jet of gas pushes rock left:
|..@....|
|.@@...|
|..@....|
|.......|
|..####.|
+-------+

Rock falls 1 unit:
|..@....|
|.@@...|
|..@....|
|..####.|
+-------+

Jet of gas pushes rock right:
|...@...|
|..@@..|
|...@...|
|..####.|
+-------+

Rock falls 1 unit, causing it to come to rest:
|...#...|
|..###..|
|...#...|
|..####.|
+-------+

A new rock begins falling:
|....@..|
|....@..|
|..@@..|
|.......|
|.......|
|.......|
|...#...|
|..###..|
|...#...|
|..####.|
+-------+
}\if{html}{\out{</div>}}

The moment each of the next few rocks begins falling, you would see
this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{|..@....|
|..@....|
|..@....|
|..@....|
|.......|
|.......|
|.......|
|..#....|
|..#....|
|####...|
|..###..|
|...#...|
|..####.|
+-------+

|..@...|
|..@...|
|.......|
|.......|
|.......|
|....#..|
|..#.#..|
|..#.#..|
|#####..|
|..###..|
|...#...|
|..####.|
+-------+

|..@@.|
|.......|
|.......|
|.......|
|....##.|
|....##.|
|....#..|
|..#.#..|
|..#.#..|
|#####..|
|..###..|
|...#...|
|..####.|
+-------+

|...@...|
|..@@..|
|...@...|
|.......|
|.......|
|.......|
|.####..|
|....##.|
|....##.|
|....#..|
|..#.#..|
|..#.#..|
|#####..|
|..###..|
|...#...|
|..####.|
+-------+

|....@..|
|....@..|
|..@@..|
|.......|
|.......|
|.......|
|..#....|
|.###...|
|..#....|
|.####..|
|....##.|
|....##.|
|....#..|
|..#.#..|
|..#.#..|
|#####..|
|..###..|
|...#...|
|..####.|
+-------+

|..@....|
|..@....|
|..@....|
|..@....|
|.......|
|.......|
|.......|
|.....#.|
|.....#.|
|..####.|
|.###...|
|..#....|
|.####..|
|....##.|
|....##.|
|....#..|
|..#.#..|
|..#.#..|
|#####..|
|..###..|
|...#...|
|..####.|
+-------+

|..@...|
|..@...|
|.......|
|.......|
|.......|
|....#..|
|....#..|
|....##.|
|....##.|
|..####.|
|.###...|
|..#....|
|.####..|
|....##.|
|....##.|
|....#..|
|..#.#..|
|..#.#..|
|#####..|
|..###..|
|...#...|
|..####.|
+-------+

|..@@.|
|.......|
|.......|
|.......|
|....#..|
|....#..|
|....##.|
|##..##.|
|######.|
|.###...|
|..#....|
|.####..|
|....##.|
|....##.|
|....#..|
|..#.#..|
|..#.#..|
|#####..|
|..###..|
|...#...|
|..####.|
+-------+
}\if{html}{\out{</div>}}

To prove to the elephants your simulation is accurate, they want to know
how tall the tower will get after 2022 rocks have stopped (but before
the 2023rd rock begins falling). In this example, the tower of rocks
will be \emph{\code{3068}} units tall.

\emph{How many units tall will the tower of rocks be after 2022 rocks have
stopped falling?}

\strong{Part Two}

The elephants are not impressed by your simulation. They demand to know
how tall the tower will be after \emph{\code{1000000000000}} rocks have stopped!
Only then will they feel confident enough to proceed through the cave.

In the example above, the tower would be \emph{\code{1514285714288}} units tall!

\emph{How tall will the tower be after \code{1000000000000} rocks have stopped?}
}
\examples{
f17a(example_data_17())
f17b()
The tall, vertical chamber is exactly *seven units wide*. Each rock
appears so that its left edge is two units away from the left wall and
its bottom edge is three units above the highest rock in the room (or
the floor, if there isn\'t one).
}
