% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day18.R
\name{day18}
\alias{day18}
\alias{f18a}
\alias{f18b}
\alias{example_data_18}
\title{Day 18: Boiling Boulders}
\usage{
f18a(x)

f18b(x)

example_data_18(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f18a(x)} returns .... For Part Two,
\code{f18b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2022/day/18}{Boiling Boulders}
}
\details{
\strong{Part One}

You and the elephants finally reach fresh air. You\'ve emerged near the
base of a large volcano that seems to be actively erupting! Fortunately,
the lava seems to be flowing away from you and toward the ocean.

Bits of lava are still being ejected toward you, so you\'re sheltering
in the cavern exit a little longer. Outside the cave, you can see the
lava landing in a pond and hear it loudly hissing as it solidifies.

Depending on the specific compounds in the lava and speed at which it
cools, it might be forming
\href{https://en.wikipedia.org/wiki/Obsidian}{obsidian}{target="_blank"}! The
cooling rate should be based on the surface area of the lava droplets,
so you take a quick scan of a droplet as it flies past you (your puzzle
input).

Because of how quickly the lava is moving, the scan isn\'t very good;
its resolution is quite low and, as a result, it approximates the shape
of the lava droplet with \emph{1x1x1
[cubes]{title="Unfortunately, you forgot your flint and steel in another dimension."}
on a 3D grid}, each given as its \verb{x,y,z} position.

To approximate the surface area, count the number of sides of each cube
that are not immediately connected to another cube. So, if your scan
were only two adjacent cubes like \verb{1,1,1} and \verb{2,1,1}, each cube would
have a single side covered and five sides exposed, a total surface area
of \emph{\code{10}} sides.

Here\'s a larger example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{2,2,2
1,2,2
3,2,2
2,1,2
2,3,2
2,2,1
2,2,3
2,2,4
2,2,6
1,2,5
3,2,5
2,1,5
2,3,5
}\if{html}{\out{</div>}}

In the above example, after counting up all the sides that aren\'t
connected to another cube, the total surface area is \emph{\code{64}}.

\emph{What is the surface area of your scanned lava droplet?}

\strong{Part Two}

Something seems off about your calculation. The cooling rate depends on
exterior surface area, but your calculation also included the surface
area of air pockets trapped in the lava droplet.

Instead, consider only cube sides that could be reached by the water and
steam as the lava droplet tumbles into the pond. The steam will expand
to reach as much as possible, completely displacing any air on the
outside of the lava droplet but never expanding diagonally.

In the larger example above, exactly one cube of air is trapped within
the lava droplet (at \verb{2,2,5}), so the exterior surface area of the lava
droplet is \emph{\code{58}}.

\emph{What is the exterior surface area of your scanned lava droplet?}
}
\examples{
f18a(example_data_18())
f18b()
}
