% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day19.R
\name{day19}
\alias{day19}
\alias{f19a}
\alias{f19b}
\alias{example_data_19}
\title{Day 19: Aplenty}
\usage{
f19a(x)

f19b(x)

example_data_19(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f19a(x)} returns .... For Part Two,
\code{f19b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2023/day/19}{Aplenty}
}
\details{
\strong{Part One}

The Elves of Gear Island are thankful for your help and send you on your
way. They even have a hang glider that someone \href{9}{stole} from Desert
Island; since you\'re already going that direction, it would help them a
lot if you would use it to get down there and return it to them.

As you reach the bottom of the \emph{relentless avalanche of machine parts},
you discover that they\'re already forming a formidable heap. Don\'t
worry, though - a group of Elves is already here organizing the parts,
and they have a
[\emph{system}]{title="This part sparks joy. This part sparks joy. This part ALSO sparks joy... I think we need a different system."}.

To start, each part is rated in each of four categories:
\itemize{
\item \code{x}: E\emph{x}tremely cool looking
\item \code{m}: \emph{M}usical (it makes a noise when you hit it)
\item \code{a}: \emph{A}erodynamic
\item \code{s}: \emph{S}hiny
}

Then, each part is sent through a series of \emph{workflows} that will
ultimately \emph{accept} or \emph{reject} the part. Each workflow has a name and
contains a list of \emph{rules}; each rule specifies a condition and where to
send the part if the condition is true. The first rule that matches the
part being considered is applied immediately, and the part moves on to
the destination described by the rule. (The last rule in each workflow
has no condition and always applies if reached.)

Consider the workflow \verb{ex\{x>10:one,m<20:two,a>30:R,A\}}. This workflow is
named \code{ex} and contains four rules. If workflow \code{ex} were considering a
specific part, it would perform the following steps in order:
\itemize{
\item Rule \"\code{x>10:one}\": If the part\'s \code{x} is more than \code{10}, send the
part to the workflow named \code{one}.
\item Rule \"\code{m<20:two}\": Otherwise, if the part\'s \code{m} is less than
\code{20}, send the part to the workflow named \code{two}.
\item Rule \"\code{a>30:R}\": Otherwise, if the part\'s \code{a} is more than \code{30},
the part is immediately \emph{rejected} (\code{R}).
\item Rule \"\code{A}\": Otherwise, because no other rules matched the part,
the part is immediately \emph{accepted} (\code{A}).
}

If a part is sent to another workflow, it immediately switches to the
start of that workflow instead and never returns. If a part is
\emph{accepted} (sent to \code{A}) or \emph{rejected} (sent to \code{R}), the part
immediately stops any further processing.

The system works, but it\'s not keeping up with the torrent of weird
metal shapes. The Elves ask if you can help sort a few parts and give
you the list of workflows and some part ratings (your puzzle input). For
example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{px\{a2090:A,rfg\}
pv\{a>1716:R,A\}
lnx\{m>1548:A,A\}
rfg\{s2440:R,A\}
qs\{s>3448:A,lnx\}
qkq\{x2662:A,R\}
in\{s2770:qs,m3333:R,R\}
hdj\{m>838:A,pv\}

\{x=787,m=2655,a=1222,s=2876\}
\{x=1679,m=44,a=2067,s=496\}
\{x=2036,m=264,a=79,s=2244\}
\{x=2461,m=1339,a=466,s=291\}
\{x=2127,m=1623,a=2188,s=1013\}
}\if{html}{\out{</div>}}

The workflows are listed first, followed by a blank line, then the
ratings of the parts the Elves would like you to sort. All parts begin
in the workflow named \verb{in}. In this example, the five listed parts go
through the following workflows:
\itemize{
\item \verb{\{x=787,m=2655,a=1222,s=2876\}}: \verb{in} -\> \code{qqz} -\> \code{qs} -\> \code{lnx}
-\> \emph{\code{A}}
\item \verb{\{x=1679,m=44,a=2067,s=496\}}: \verb{in} -\> \code{px} -\> \code{rfg} -\> \code{gd} -\>
\emph{\code{R}}
\item \verb{\{x=2036,m=264,a=79,s=2244\}}: \verb{in} -\> \code{qqz} -\> \code{hdj} -\> \code{pv} -\>
\emph{\code{A}}
\item \verb{\{x=2461,m=1339,a=466,s=291\}}: \verb{in} -\> \code{px} -\> \code{qkq} -\> \code{crn} -\>
\emph{\code{R}}
\item \verb{\{x=2127,m=1623,a=2188,s=1013\}}: \verb{in} -\> \code{px} -\> \code{rfg} -\> \emph{\code{A}}
}

Ultimately, three parts are \emph{accepted}. Adding up the \code{x}, \code{m}, \code{a}, and
\code{s} rating for each of the accepted parts gives \code{7540} for the part with
\code{x=787}, \code{4623} for the part with \code{x=2036}, and \code{6951} for the part with
\code{x=2127}. Adding all of the ratings for \emph{all} of the accepted parts
gives the sum total of \emph{\code{19114}}.

Sort through all of the parts you\'ve been given; \emph{what do you get if
you add together all of the rating numbers for all of the parts that
ultimately get accepted?}

\strong{Part Two}

Even with your help, the sorting process \emph{still} isn\'t fast enough.

One of the Elves comes up with a new plan: rather than sort parts
individually through all of these workflows, maybe you can figure out in
advance which combinations of ratings will be accepted or rejected.

Each of the four ratings (\code{x}, \code{m}, \code{a}, \code{s}) can have an integer value
ranging from a minimum of \code{1} to a maximum of \code{4000}. Of \emph{all possible
distinct combinations} of ratings, your job is to figure out which ones
will be \emph{accepted}.

In the above example, there are \emph{\code{167409079868000}} distinct
combinations of ratings that will be accepted.

Consider only your list of workflows; the list of part ratings that the
Elves wanted you to sort is no longer relevant. \emph{How many distinct
combinations of ratings will be accepted by the Elves\' workflows?}
}
\examples{
f19a(example_data_19())
f19b()
}
