% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day21.R
\name{day21}
\alias{day21}
\alias{f21a}
\alias{f21b}
\alias{example_data_21}
\title{Day 21: Monkey Math}
\usage{
f21a(x)

f21b(x)

example_data_21(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f21a(x)} returns .... For Part Two,
\code{f21b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2022/day/21}{Monkey Math}
}
\details{
\strong{Part One}

The \href{11}{monkeys} are back! You\'re worried they\'re going to try to
steal your stuff again, but it seems like they\'re just holding their
ground and making various monkey noises at you.

Eventually, one of the elephants realizes you don\'t speak monkey and
comes over to interpret. As it turns out, they overheard you talking
about trying to find the grove; they can show you a shortcut if you
answer their \emph{riddle}.

Each monkey is given a \emph{job}: either to \emph{yell a specific number} or to
\emph{yell the result of a math operation}. All of the number-yelling monkeys
know their number from the start; however, the math operation monkeys
need to wait for two other monkeys to yell a number, and those two other
monkeys might \emph{also} be waiting on other monkeys.

Your job is to \emph{work out the number the monkey named \code{root} will yell}
before the monkeys figure it out themselves.

For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{root: pppw + sjmn
dbpl: 5
cczh: sllz + lgvd
zczc: 2
ptdq: humn - dvpt
dvpt: 3
lfqf: 4
humn: 5
ljgn: 2
sjmn: drzm * dbpl
sllz: 4
pppw: cczh / lfqf
lgvd: ljgn * ptdq
drzm: hmdt - zczc
hmdt: 32
}\if{html}{\out{</div>}}

Each line contains the name of a monkey, a colon, and then the job of
that monkey:
\itemize{
\item A lone number means the monkey\'s job is simply to yell that number.
\item A job like \code{aaaa + bbbb} means the monkey waits for monkeys \code{aaaa}
and \code{bbbb} to yell each of their numbers; the monkey then yells the
sum of those two numbers.
\item \code{aaaa - bbbb} means the monkey yells \code{aaaa}\'s number minus
\code{bbbb}\'s number.
\item Job \code{aaaa * bbbb} will yell \code{aaaa}\'s number multiplied by \code{bbbb}\'s
number.
\item Job \code{aaaa / bbbb} will yell \code{aaaa}\'s number divided by \code{bbbb}\'s
number.
}

So, in the above example, monkey \code{drzm} has to wait for monkeys \code{hmdt}
and \code{zczc} to yell their numbers. Fortunately, both \code{hmdt} and \code{zczc}
have jobs that involve simply yelling a single number, so they do this
immediately: \code{32} and \code{2}. Monkey \code{drzm} can then yell its number by
finding \code{32} minus \code{2}: \emph{\code{30}}.

Then, monkey \code{sjmn} has one of its numbers (\code{30}, from monkey \code{drzm}),
and already has its other number, \code{5}, from \code{dbpl}. This allows it to
yell its own number by finding \code{30} multiplied by \code{5}: \emph{\code{150}}.

This process continues until \code{root} yells a number: \emph{\code{152}}.

However, your actual situation involves [considerably more
monkeys]{title="Advent of Code 2022: Now With Considerably More Monkeys"}.
\emph{What number will the monkey named \code{root} yell?}

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
f21a(example_data_21())
f21b()
}
