% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day23.R
\name{day23}
\alias{day23}
\alias{f23a}
\alias{f23b}
\alias{example_data_23}
\title{Day 23: Unstable Diffusion}
\usage{
f23a(x)

f23b(x)

example_data_23(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f23a(x)} returns .... For Part Two,
\code{f23b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2022/day/23}{Unstable Diffusion}
}
\details{
\strong{Part One}

You enter a large crater of gray dirt where the grove is supposed to be.
All around you, plants you imagine were expected to be full of fruit are
instead withered and broken. A large group of Elves has formed in the
middle of the grove.

\"\...but this volcano has been dormant for months. Without ash, the
fruit can\'t grow!\"

You look up to see a massive, snow-capped mountain towering above you.

\"It\'s not like there are other active volcanoes here; we\'ve looked
everywhere.\"

\"But our scanners show active magma flows; clearly it\'s going
\emph{somewhere}.\"

They finally notice you at the edge of the grove, your pack almost
overflowing from the random \emph{star} fruit you\'ve been collecting. Behind
you, elephants and monkeys explore the grove, looking concerned. Then,
the Elves recognize the ash cloud slowly spreading above your recent
detour.

\"Why do you\--\" \"How is\--\" \"Did you just\--\"

Before any of them can form a complete question, another Elf speaks up:
\"Okay, new plan. We have almost enough fruit already, and ash from the
plume should spread here eventually. If we quickly plant new seedlings
now, we can still make it to the extraction point. Spread out!\"

The Elves each reach into their pack and pull out a tiny plant. The
plants rely on important nutrients from the ash, so they can\'t be
planted too close together.

There isn\'t enough time to let the Elves figure out where to plant the
seedlings themselves; you quickly scan the grove (your puzzle input) and
note their positions.

For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{....#..
..###.#
#...#.#
.#...##
#.###..
##.#.##
.#..#..
}\if{html}{\out{</div>}}

The scan shows Elves \verb{#} and empty ground \code{.}; outside your scan, more
empty ground extends a long way in every direction. The scan is oriented
so that \emph{north is up}; orthogonal directions are written N (north), S
(south), W (west), and E (east), while diagonal directions are written
NE, NW, SE, SW.

The Elves follow a time-consuming process to figure out where they
should each go; you can speed up this process considerably. The process
consists of some number of \emph{rounds} during which Elves alternate between
considering where to move and actually moving.

During the \emph{first half} of each round, each Elf considers the eight
positions adjacent to themself. If no other Elves are in one of those
eight positions, the Elf \emph{does not do anything} during this round.
Otherwise, the Elf looks in each of four directions in the following
order and \emph{proposes} moving one step in the \emph{first valid direction}:
\itemize{
\item If there is no Elf in the N, NE, or NW adjacent positions, the Elf
proposes moving \emph{north} one step.
\item If there is no Elf in the S, SE, or SW adjacent positions, the Elf
proposes moving \emph{south} one step.
\item If there is no Elf in the W, NW, or SW adjacent positions, the Elf
proposes moving \emph{west} one step.
\item If there is no Elf in the E, NE, or SE adjacent positions, the Elf
proposes moving \emph{east} one step.
}

After each Elf has had a chance to propose a move, the \emph{second half} of
the round can begin. Simultaneously, each Elf moves to their proposed
destination tile if they were the \emph{only} Elf to propose moving to that
position. If two or more Elves propose moving to the same position,
\emph{none} of those Elves move.

Finally, at the end of the round, the \emph{first direction} the Elves
considered is moved to the end of the list of directions. For example,
during the second round, the Elves would try proposing a move to the
south first, then west, then east, then north. On the third round, the
Elves would first consider west, then east, then north, then south.

As a smaller example, consider just these five Elves:

\if{html}{\out{<div class="sourceCode">}}\preformatted{.....
..##.
..#..
.....
..##.
.....
}\if{html}{\out{</div>}}

The northernmost two Elves and southernmost two Elves all propose moving
north, while the middle Elf cannot move north and proposes moving south.
The middle Elf proposes the same destination as the southwest Elf, so
neither of them move, but the other three do:

\if{html}{\out{<div class="sourceCode">}}\preformatted{..##.
.....
..#..
...#.
..#..
.....
}\if{html}{\out{</div>}}

Next, the northernmost two Elves and the southernmost Elf all propose
moving south. Of the remaining middle two Elves, the west one cannot
move south and proposes moving west, while the east one cannot move
south \emph{or} west and proposes moving east. All five Elves succeed in
moving to their proposed positions:

\if{html}{\out{<div class="sourceCode">}}\preformatted{.....
..##.
.#...
....#
.....
..#..
}\if{html}{\out{</div>}}

Finally, the southernmost two Elves choose not to move at all. Of the
remaining three Elves, the west one proposes moving west, the east one
proposes moving east, and the middle one proposes moving north; all
three succeed in moving:

\if{html}{\out{<div class="sourceCode">}}\preformatted{..#..
....#
#....
....#
.....
..#..
}\if{html}{\out{</div>}}

At this point, no Elves need to move, and so the process ends.

The larger example above proceeds as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{== Initial State ==
..............
..............
.......#......
.....###.#....
...#...#.#....
....#...##....
...#.###......
...##.#.##....
....#..#......
..............
..............
..............

== End of Round 1 ==
..............
.......#......
.....#...#....
...#..#.#.....
.......#..#...
....#.#.##....
..#..#.#......
..#.#.#.##....
..............
....#..#......
..............
..............

== End of Round 2 ==
..............
.......#......
....#.....#...
...#..#.#.....
.......#...#..
...#..#.#.....
.#...#.#.#....
..............
..#.#.#.##....
....#..#......
..............
..............

== End of Round 3 ==
..............
.......#......
.....#....#...
..#..#...#....
.......#...#..
...#..#.#.....
.#..#.....#...
.......##.....
..##.#....#...
...#..........
.......#......
..............

== End of Round 4 ==
..............
.......#......
......#....#..
..#...##......
...#.....#.#..
.........#....
.#...###..#...
..#......#....
....##....#...
....#.........
.......#......
..............

== End of Round 5 ==
.......#......
..............
..#..#.....#..
.........#....
......##...#..
.#.#.####.....
...........#..
....##..#.....
..#...........
..........#...
....#..#......
..............
}\if{html}{\out{</div>}}

After a few more rounds\...

\if{html}{\out{<div class="sourceCode">}}\preformatted{== End of Round 10 ==
.......#......
...........#..
..#.#..#......
......#.......
...#.....#..#.
.#......##....
.....##.......
..#........#..
....#.#..#....
..............
....#..#..#...
..............
}\if{html}{\out{</div>}}

To make sure they\'re on the right track, the Elves like to check after
round 10 that they\'re making good progress toward covering enough
ground. To do this, count the number of empty ground tiles contained by
the smallest rectangle that contains every Elf. (The edges of the
rectangle should be aligned to the N/S/E/W directions; the Elves do not
have the patience to calculate [arbitrary
rectangles]{title="Arbitrary Rectangles is my Piet Mondrian cover band."}.)
In the above example, that rectangle is:

\if{html}{\out{<div class="sourceCode">}}\preformatted{......#.....
..........#.
.#.#..#.....
.....#......
..#.....#..#
#......##...
....##......
.#........#.
...#.#..#...
............
...#..#..#..
}\if{html}{\out{</div>}}

In this region, the number of empty ground tiles is \emph{\code{110}}.

Simulate the Elves\' process and find the smallest rectangle that
contains the Elves after 10 rounds. \emph{How many empty ground tiles does
that rectangle contain?}

\strong{Part Two}

It seems you\'re on the right track. Finish simulating the process and
figure out where the Elves need to go. How many rounds did you save
them?

In the example above, the \emph{first round where no Elf moved} was round
\emph{\code{20}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{.......#......
....#......#..
..#.....#.....
......#.......
...#....#.#..#
#.............
....#.....#...
..#.....#.....
....#.#....#..
.........#....
....#......#..
.......#......
}\if{html}{\out{</div>}}

Figure out where the Elves need to go. \emph{What is the number of the first
round where no Elf moves?}
}
\examples{
f23a(example_data_23())
f23b()
}
