% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day24.R
\name{day24}
\alias{day24}
\alias{f24a}
\alias{f24b}
\alias{example_data_24}
\title{Day 24: Never Tell Me The Odds}
\usage{
f24a(x, pt1, pt2)

f24b(x)

example_data_24(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f24a(x)} returns .... For Part Two,
\code{f24b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2023/day/24}{Never Tell Me The Odds}
}
\details{
\strong{Part One}

It seems like something is going wrong with the snow-making process.
Instead of forming snow, the water that\'s been absorbed into the air
seems to be forming
\href{https://en.wikipedia.org/wiki/Hail}{hail}{target="_blank"}!

Maybe there\'s something you can do to break up the hailstones?

Due to strong, probably-magical winds, the hailstones are all flying
through the air in perfectly linear trajectories. You make a note of
each hailstone\'s \emph{position} and \emph{velocity} (your puzzle input). For
example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{19, 13, 30 @ -2,  1, -2
18, 19, 22 @ -1, -1, -2
20, 25, 34 @ -2, -2, -4
12, 31, 28 @ -1, -2, -1
20, 19, 15 @  1, -5, -3
}\if{html}{\out{</div>}}

Each line of text corresponds to the position and velocity of a single
hailstone. The positions indicate where the hailstones are \emph{right now}
(at time \code{0}). The velocities are constant and indicate exactly how far
each hailstone will move in \emph{one nanosecond}.

Each line of text uses the format \verb{px py pz @ vx vy vz}. For instance,
the hailstone specified by \verb{20, 19, 15 @ 1, -5, -3} has initial X
position \code{20}, Y position \code{19}, Z position \code{15}, X velocity \code{1}, Y
velocity \code{-5}, and Z velocity \code{-3}. After one nanosecond, the hailstone
would be at \verb{21, 14, 12}.

Perhaps you won\'t have to do anything. How likely are the hailstones to
collide with each other and smash into tiny ice crystals?

To estimate this, consider only the X and Y axes; \emph{ignore the Z axis}.
Looking \emph{forward in time}, how many of the hailstones\' \emph{paths} will
intersect within a test area? (The hailstones themselves don\'t have to
collide, just test for intersections between the paths they will trace.)

In this example, look for intersections that happen with an X and Y
position each at least \code{7} and at most \code{27}; in your actual data,
you\'ll need to check a much larger test area. Comparing all pairs of
hailstones\' future paths produces the following results:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Hailstone A: 19, 13, 30 @ -2, 1, -2
Hailstone B: 18, 19, 22 @ -1, -1, -2
Hailstones' paths will cross inside the test area (at x=14.333, y=15.333).

Hailstone A: 19, 13, 30 @ -2, 1, -2
Hailstone B: 20, 25, 34 @ -2, -2, -4
Hailstones' paths will cross inside the test area (at x=11.667, y=16.667).

Hailstone A: 19, 13, 30 @ -2, 1, -2
Hailstone B: 12, 31, 28 @ -1, -2, -1
Hailstones' paths will cross outside the test area (at x=6.2, y=19.4).

Hailstone A: 19, 13, 30 @ -2, 1, -2
Hailstone B: 20, 19, 15 @ 1, -5, -3
Hailstones' paths crossed in the past for hailstone A.

Hailstone A: 18, 19, 22 @ -1, -1, -2
Hailstone B: 20, 25, 34 @ -2, -2, -4
Hailstones' paths are parallel; they never intersect.

Hailstone A: 18, 19, 22 @ -1, -1, -2
Hailstone B: 12, 31, 28 @ -1, -2, -1
Hailstones' paths will cross outside the test area (at x=-6, y=-5).

Hailstone A: 18, 19, 22 @ -1, -1, -2
Hailstone B: 20, 19, 15 @ 1, -5, -3
Hailstones' paths crossed in the past for both hailstones.

Hailstone A: 20, 25, 34 @ -2, -2, -4
Hailstone B: 12, 31, 28 @ -1, -2, -1
Hailstones' paths will cross outside the test area (at x=-2, y=3).

Hailstone A: 20, 25, 34 @ -2, -2, -4
Hailstone B: 20, 19, 15 @ 1, -5, -3
Hailstones' paths crossed in the past for hailstone B.

Hailstone A: 12, 31, 28 @ -1, -2, -1
Hailstone B: 20, 19, 15 @ 1, -5, -3
Hailstones' paths crossed in the past for both hailstones.
}\if{html}{\out{</div>}}

So, in this example, \emph{\code{2}} hailstones\' future paths cross inside the
boundaries of the test area.

However, you\'ll need to search a much larger test area if you want to
see if any hailstones might collide. Look for intersections that happen
with an X and Y position each at least \code{200000000000000} and at most
\code{400000000000000}. Disregard the Z axis entirely.

Considering only the X and Y axes, check all pairs of hailstones\'
future paths for intersections. \emph{How many of these intersections occur
within the test area?}

\strong{Part Two}

Upon further analysis, it doesn\'t seem like \emph{any} hailstones will
naturally collide. It\'s up to you to fix that!

You find a rock on the ground nearby. While it seems extremely unlikely,
if you throw it just right, you should be able to \emph{hit every hailstone
in a single throw}!

You can use the probably-magical winds to reach \emph{any integer position}
you like and to propel the rock at \emph{any integer velocity}. Now
\emph{including the Z axis} in your calculations, if you throw the rock at
time \code{0}, where do you need to be so that the rock \emph{perfectly collides
with every hailstone}? Due to [probably-magical
inertia]{title="What, you've never studied probably-magical physics?"},
the rock won\'t slow down or change direction when it collides with a
hailstone.

In the example above, you can achieve this by moving to position
\verb{24, 13, 10} and throwing the rock at velocity \verb{-3, 1, 2}. If you do
this, you will hit every hailstone as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Hailstone: 19, 13, 30 @ -2, 1, -2
Collision time: 5
Collision position: 9, 18, 20

Hailstone: 18, 19, 22 @ -1, -1, -2
Collision time: 3
Collision position: 15, 16, 16

Hailstone: 20, 25, 34 @ -2, -2, -4
Collision time: 4
Collision position: 12, 17, 18

Hailstone: 12, 31, 28 @ -1, -2, -1
Collision time: 6
Collision position: 6, 19, 22

Hailstone: 20, 19, 15 @ 1, -5, -3
Collision time: 1
Collision position: 21, 14, 12
}\if{html}{\out{</div>}}

Above, each hailstone is identified by its initial position and its
velocity. Then, the time and position of that hailstone\'s collision
with your rock are given.

After 1 nanosecond, the rock has \emph{exactly the same position} as one of
the hailstones, obliterating it into ice dust! Another hailstone is
smashed to bits two nanoseconds after that. After a total of 6
nanoseconds, all of the hailstones have been destroyed.

So, at time \code{0}, the rock needs to be at X position \code{24}, Y position
\code{13}, and Z position \code{10}. Adding these three coordinates together
produces \emph{\code{47}}. (Don\'t add any coordinates from the rock\'s velocity.)

Determine the exact position and velocity the rock needs to have at time
\code{0} so that it perfectly collides with every hailstone. \emph{What do you get
if you add up the X, Y, and Z coordinates of that initial position?}
}
\examples{
f24a(example_data_24())
f24b()
}
