% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day25.R
\name{day25}
\alias{day25}
\alias{f25a}
\alias{f25b}
\alias{example_data_25}
\title{Day 25: Snowverload}
\usage{
f25a(x)

f25b(x)

example_data_25(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f25a(x)} returns .... For Part Two,
\code{f25b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2023/day/25}{Snowverload}
}
\details{
\strong{Part One}

\emph{Still} somehow without snow, you go to the last place you haven\'t
checked: the center of Snow Island, directly below the waterfall.

Here, someone has clearly been trying to fix the problem. Scattered
everywhere are hundreds of weather machines, almanacs, communication
modules, hoof prints, machine parts, mirrors, lenses, and so on.

Somehow, everything has been \emph{wired together} into a massive
snow-producing apparatus, but nothing seems to be running. You check a
tiny screen on one of the communication modules: \verb{Error 2023}. It
doesn\'t say what \verb{Error 2023} means, but it \emph{does} have the phone
number for a support line printed on it.

\"Hi, you\'ve reached Weather Machines And So On, Inc. How can I help
you?\" You explain the situation.

\"Error 2023, you say? Why, that\'s a power overload error, of course!
It means you have too many components plugged in. Try unplugging some
components and\--\" You explain that there are hundreds of components
here and you\'re in a bit of a hurry.

\"Well, let\'s see how bad it is; do you see a \emph{big red reset button}
somewhere? It should be on its own module. If you push it, it probably
won\'t fix anything, but it\'ll report how overloaded things are.\"
After a minute or two, you find the reset button; it\'s so big that it
takes two hands just to get enough leverage to push it. Its screen then
displays:

\if{html}{\out{<div class="sourceCode">}}\preformatted{SYSTEM OVERLOAD!

Connected components would require
power equal to at least 100 stars!
}\if{html}{\out{</div>}}

\"Wait, \emph{how} many components did you say are plugged in? With that much
equipment, you could produce snow for an \emph{entire}\--\" You disconnect
the call.

You have nowhere near that many stars - you need to find a way to
disconnect at least half of the equipment here, but it\'s already
Christmas! You only have time to disconnect \emph{three wires}.

Fortunately, someone left a wiring diagram (your puzzle input) that
shows \emph{how the components are connected}. For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{jqt: rhn xhk nvd
rsh: frs pzl lsr
xhk: hfx
cmg: qnr nvd lhk bvb
rhn: xhk bvb hfx
bvb: xhk hfx
pzl: lsr hfx nvd
qnr: nvd
ntq: jqt hfx bvb xhk
nvd: lhk
lsr: lhk
rzs: qnr cmg lsr rsh
frs: qnr lhk lsr
}\if{html}{\out{</div>}}

Each line shows the \emph{name of a component}, a colon, and then \emph{a list of
other components} to which that component is connected. Connections
aren\'t directional; \code{abc: xyz} and \code{xyz: abc} both represent the same
configuration. Each connection between two components is represented
only once, so some components might only ever appear on the left or
right side of a colon.

In this example, if you disconnect the wire between \code{hfx}/\code{pzl}, the
wire between \code{bvb}/\code{cmg}, and the wire between \code{nvd}/\code{jqt}, you will
\emph{divide the components into two separate, disconnected groups}:
\itemize{
\item \emph{\code{9}} components: \code{cmg}, \code{frs}, \code{lhk}, \code{lsr}, \code{nvd}, \code{pzl}, \code{qnr},
\code{rsh}, and \code{rzs}.
\item \emph{\code{6}} components: \code{bvb}, \code{hfx}, \code{jqt}, \code{ntq}, \code{rhn}, and \code{xhk}.
}

Multiplying the sizes of these groups together produces \emph{\code{54}}.

Find the three wires you need to disconnect in order to divide the
components into two separate groups. \emph{What do you get if you multiply
the sizes of these two groups together?}

\strong{Part Two}

You climb over weather machines, under giant springs, and narrowly avoid
a pile of pipes as you find and disconnect the three wires.

A moment after you disconnect the last wire, the big red reset button
module makes a small ding noise:

\if{html}{\out{<div class="sourceCode">}}\preformatted{System overload resolved!
Power required is now 50 stars.
}\if{html}{\out{</div>}}

Out of the corner of your eye, you notice goggles and a loose-fitting
hard hat peeking at you from behind an ultra crucible. You think you see
a [faint glow]{title="i help"}, but before you can investigate, you hear
another small ding:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Power required is now 49 stars.

Please supply the necessary stars and
push the button to restart the system.
}\if{html}{\out{</div>}}
}
\examples{
f25a(example_data_25())
f25b()
}
