% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day20.R
\name{makefun}
\alias{makefun}
\title{\emph{Conjunction} modules (prefix \code{&}) \emph{remember} the type of the most
recent pulse received from \emph{each} of their connected input modules; they
initially default to remembering a \emph{low pulse} for each input. When a
pulse is received, the conjunction module first updates its memory for
that input. Then, if it remembers \emph{high pulses} for all inputs, it sends
a \emph{low pulse}; otherwise, it sends a \emph{high pulse}
\emph{Flip-flop} modules (prefix \verb{\%}) are either \emph{on} or \emph{off}; they are
initially \emph{off}. If a flip-flop module receives a high pulse, it is
ignored and nothing happens. However, if a flip-flop module receives a
low pulse, it \emph{flips between on and off}. If it was off, it turns on and
sends a high pulse. If it was on, it turns off and sends a low pulse.}
\usage{
makefun(x, s)
}
\description{
\emph{Conjunction} modules (prefix \code{&}) \emph{remember} the type of the most
recent pulse received from \emph{each} of their connected input modules; they
initially default to remembering a \emph{low pulse} for each input. When a
pulse is received, the conjunction module first updates its memory for
that input. Then, if it remembers \emph{high pulses} for all inputs, it sends
a \emph{low pulse}; otherwise, it sends a \emph{high pulse}
\emph{Flip-flop} modules (prefix \verb{\%}) are either \emph{on} or \emph{off}; they are
initially \emph{off}. If a flip-flop module receives a high pulse, it is
ignored and nothing happens. However, if a flip-flop module receives a
low pulse, it \emph{flips between on and off}. If it was off, it turns on and
sends a high pulse. If it was on, it turns off and sends a low pulse.
}
