% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day17.R
\docType{data}
\name{rocks}
\alias{rocks}
\title{\subsection{}{
}}
\format{
An object of class \code{list} of length 5.
}
\usage{
rocks
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{.#.
###
.#.
}\if{html}{\out{</div>}}
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{..#
..#
###

#
#
#
#

##
##
}\if{html}{\out{</div>}}
}
\keyword{datasets}
