% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day22.R
\name{wrap}
\alias{wrap}
\title{11112222
11112222
11112222
11112222
3333
3333
3333
3333
44445555
44445555
44445555
44445555
6666
6666
6666
6666}
\usage{
wrap(pos, facing)
}
\description{
11112222
11112222
11112222
11112222
3333
3333
3333
3333
44445555
44445555
44445555
44445555
6666
6666
6666
6666
}
