% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day01.R
\name{day01}
\alias{day01}
\alias{f01a}
\alias{f01b}
\alias{example_data_01}
\title{Day 01: Secret Entrance}
\usage{
f01a(x)

f01b(x)

example_data_01(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f01a(x)} returns .... For Part Two,
\code{f01b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2025/day/1}{Secret Entrance}
}
\details{
\strong{Part One}

The Elves have good news and bad news.

The good news is that they\'ve discovered \href{https://en.wikipedia.org/wiki/Project_management}{project management}{target="_blank"}!
This has given them the tools they need to prevent their usual Christmas
emergency. For example, they now know that the North Pole decorations
need to be finished soon so that other critical tasks can start on time.

The bad news is that they\'ve realized they have a \emph{different}
emergency: according to their resource planning, none of them have any
time left to decorate the North Pole!

To save Christmas, the Elves need \emph{you} to \emph{finish decorating the North
Pole by December 12th}.

Collect stars by solving puzzles. Two puzzles will be made available on
each day; the second puzzle is unlocked when you complete the first.
Each puzzle grants \emph{one star}. Good luck!

You arrive at the secret entrance to the North Pole base ready to start
decorating. Unfortunately, the \emph{password} seems to have been changed, so
you can\'t get in. A document taped to the wall helpfully explains:

\"Due to new security protocols, the password is locked in the safe
below. Please see the attached document for the new combination.\"

The safe has a dial with only an arrow on it; around the dial are the
numbers \code{0} through \code{99} in order. As you turn the dial, it makes a
small \emph{click} noise as it reaches each number.

The attached document (your puzzle input) contains a sequence of
\emph{rotations}, one per line, which tell you how to open the safe. A
rotation starts with an \code{L} or \code{R} which indicates whether the rotation
should be to the \emph{left} (toward lower numbers) or to the \emph{right} (toward
higher numbers). Then, the rotation has a \emph{distance} value which
indicates how many clicks the dial should be rotated in that direction.

So, if the dial were pointing at \code{11}, a rotation of \code{R8} would cause
the dial to point at \code{19}. After that, a rotation of \code{L19} would cause
it to point at \code{0}.

Because the dial is a circle, turning the dial \emph{left from \code{0}} one click
makes it point at \code{99}. Similarly, turning the dial \emph{right from \code{99}}
one click makes it point at \code{0}.

So, if the dial were pointing at \code{5}, a rotation of \code{L10} would cause it
to point at \code{95}. After that, a rotation of \code{R5} could cause it to point
at \code{0}.

The dial starts by pointing at \code{50}.

You could follow the instructions, but your recent required official
North Pole secret entrance security training seminar taught you that the
safe is actually a decoy. The actual password is \emph{the number of times
the dial is left pointing at \code{0} after any rotation in the sequence}.

For example, suppose the attached document contained the following
rotations:

\if{html}{\out{<div class="sourceCode">}}\preformatted{L68
L30
R48
L5
R60
L55
L1
L99
R14
L82
}\if{html}{\out{</div>}}

Following these rotations would cause the dial to move as follows:
\itemize{
\item The dial starts by pointing at \code{50}.
\item The dial is rotated \code{L68} to point at \code{82}.
\item The dial is rotated \code{L30} to point at \code{52}.
\item The dial is rotated \code{R48} to point at \emph{\code{0}}.
\item The dial is rotated \code{L5} to point at \code{95}.
\item The dial is rotated \code{R60} to point at \code{55}.
\item The dial is rotated \code{L55} to point at \emph{\code{0}}.
\item The dial is rotated \code{L1} to point at \code{99}.
\item The dial is rotated \code{L99} to point at \emph{\code{0}}.
\item The dial is rotated \code{R14} to point at \code{14}.
\item The dial is rotated \code{L82} to point at \code{32}.
}

Because the dial points at \code{0} a total of three times during this
process, the password in this example is \emph{\code{3}}.

Analyze the rotations in your attached document. \emph{What\'s the actual
password to open the door?}

\strong{Part Two}
You\'re sure that\'s the right password, but the door won\'t open. You
knock, but nobody answers. You build a snowman while you think.

As you\'re rolling the snowballs for your snowman, you find another
security document that must have fallen into the snow:

\"Due to newer security protocols, please use \emph{password method
[0x434C49434B]{title="You should have seen the chaos when the Elves overflowed their 32-bit password method counter."}}
until further notice.\"

You remember from the training seminar that \"method 0x434C49434B\"
means you\'re actually supposed to count the number of times \emph{any click}
causes the dial to point at \code{0}, regardless of whether it happens during
a rotation or at the end of one.

Following the same rotations as in the above example, the dial points at
zero a few extra times during its rotations:
\itemize{
\item The dial starts by pointing at \code{50}.
\item The dial is rotated \code{L68} to point at \code{82}; during this rotation, it
points at \code{0} \emph{once}.
\item The dial is rotated \code{L30} to point at \code{52}.
\item The dial is rotated \code{R48} to point at \emph{\code{0}}.
\item The dial is rotated \code{L5} to point at \code{95}.
\item The dial is rotated \code{R60} to point at \code{55}; during this rotation, it
points at \code{0} \emph{once}.
\item The dial is rotated \code{L55} to point at \emph{\code{0}}.
\item The dial is rotated \code{L1} to point at \code{99}.
\item The dial is rotated \code{L99} to point at \emph{\code{0}}.
\item The dial is rotated \code{R14} to point at \code{14}.
\item The dial is rotated \code{L82} to point at \code{32}; during this rotation, it
points at \code{0} \emph{once}.
}

In this example, the dial points at \code{0} three times at the end of a
rotation, plus three more times during a rotation. So, in this example,
the new password would be \emph{\code{6}}.

Be careful: if the dial were pointing at \code{50}, a single rotation like
\code{R1000} would cause the dial to point at \code{0} ten times before returning
back to \code{50}!

Using password method 0x434C49434B, \emph{what is the password to open the
door?}
}
\examples{
f01a(example_data_01())
f01b()
}
