% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day01.R
\name{day01}
\alias{day01}
\alias{solve01a}
\alias{solve01b}
\alias{example_data_01}
\title{Day 01: Sonar Sweep}
\usage{
solve01a(x)

solve01b(x)

example_data_01(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2021/day/1}{Sonar Sweep}
}
\examples{
solve01a(example_data_01())
solve01b(example_data_01())
}
