% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{day02}
\alias{day02}
\alias{f02a}
\alias{f02b}
\alias{example_data_02}
\title{Day 02: Gift Shop}
\usage{
f02a(x)

f02b(x)

example_data_02(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f02a(x)} returns .... For Part Two,
\code{f02b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2025/day/2}{Gift Shop}
}
\details{
\strong{Part One}

You get inside and take the elevator to its only other stop: the gift
shop. \"Thank you for visiting the North Pole!\" gleefully exclaims a
nearby sign. You aren\'t sure who is even allowed to visit the North
Pole, but you know you can access the lobby through here, and from there
you can access the rest of the North Pole base.

As you make your way through the [surprisingly
extensive]{title="They even sell lunch boxes and blue tents!"}
selection, one of the clerks recognizes you and asks for your help.

As it turns out, one of the younger Elves was playing on a gift shop
computer and managed to add a whole bunch of invalid product IDs to
their gift shop database! Surely, it would be no trouble for you to
identify the invalid product IDs for them, right?

They\'ve even checked most of the product ID ranges already; they only
have a few product ID ranges (your puzzle input) that you\'ll need to
check. For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{11-22,95-115,998-1012,1188511880-1188511890,222220-222224,
1698522-1698528,446443-446449,38593856-38593862,565653-565659,
824824821-824824827,2121212118-2121212124
}\if{html}{\out{</div>}}

(The ID ranges are wrapped here for legibility; in your input, they
appear on a single long line.)

The ranges are separated by commas (\verb{,}); each range gives its \emph{first
ID} and \emph{last ID} separated by a dash (\code{-}).

Since the young Elf was just doing silly patterns, you can find the
\emph{invalid IDs} by looking for any ID which is made only of some sequence
of digits repeated twice. So, \code{55} (\code{5} twice), \code{6464} (\code{64} twice), and
\code{123123} (\code{123} twice) would all be invalid IDs.

None of the numbers have leading zeroes; \code{0101} isn\'t an ID at all.
(\code{101} is a \emph{valid} ID that you would ignore.)

Your job is to find all of the invalid IDs that appear in the given
ranges. In the above example:
\itemize{
\item \code{11-22} has two invalid IDs, \emph{\code{11}} and \emph{\code{22}}.
\item \code{95-115} has one invalid ID, \emph{\code{99}}.
\item \code{998-1012} has one invalid ID, \emph{\code{1010}}.
\item \code{1188511880-1188511890} has one invalid ID, \emph{\code{1188511885}}.
\item \code{222220-222224} has one invalid ID, \emph{\code{222222}}.
\item \code{1698522-1698528} contains no invalid IDs.
\item \code{446443-446449} has one invalid ID, \emph{\code{446446}}.
\item \code{38593856-38593862} has one invalid ID, \emph{\code{38593859}}.
\item The rest of the ranges contain no invalid IDs.
}

Adding up all the invalid IDs in this example produces \emph{\code{1227775554}}.

\emph{What do you get if you add up all of the invalid IDs?}

\strong{Part Two}

The clerk quickly discovers that there are still invalid IDs in the
ranges in your list. Maybe the young Elf was doing other silly patterns
as well?

Now, an ID is invalid if it is made only of some sequence of digits
repeated \emph{at least} twice. So, \code{12341234} (\code{1234} two times),
\code{123123123} (\code{123} three times), \code{1212121212} (\code{12} four times), and
\code{1111111} (\code{1} seven times) are all invalid IDs.

From the same example as before:
\itemize{
\item \code{11-22} still has two invalid IDs, \emph{\code{11}} and \emph{\code{22}}.
\item \code{95-115} now has two invalid IDs, \emph{\code{99}} and \emph{\code{111}}.
\item \code{998-1012} now has two invalid IDs, \emph{\code{999}} and \emph{\code{1010}}.
\item \code{1188511880-1188511890} still has one invalid ID, \emph{\code{1188511885}}.
\item \code{222220-222224} still has one invalid ID, \emph{\code{222222}}.
\item \code{1698522-1698528} still contains no invalid IDs.
\item \code{446443-446449} still has one invalid ID, \emph{\code{446446}}.
\item \code{38593856-38593862} still has one invalid ID, \emph{\code{38593859}}.
\item \code{565653-565659} now has one invalid ID, \emph{\code{565656}}.
\item \code{824824821-824824827} now has one invalid ID, \emph{\code{824824824}}.
\item \code{2121212118-2121212124} now has one invalid ID, \emph{\code{2121212121}}.
}

Adding up all the invalid IDs in this example produces \emph{\code{4174379265}}.

\emph{What do you get if you add up all of the invalid IDs using these new
rules?}
}
\examples{
f02a(example_data_02())
f02b()
}
