% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{day02}
\alias{day02}
\alias{solve02a}
\alias{solve02b}
\alias{example_data_02}
\title{Day 02: Dive!}
\usage{
solve02a(x)

solve02b(x)

example_data_02(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2021/day/2}{Dive!}
}
\examples{
solve02a(example_data_02())
solve02b(example_data_02())
}
