% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{day02}
\alias{day02}
\alias{f02a}
\alias{f02b}
\alias{example_data_02}
\title{Day 02: Rock Paper Scissors}
\usage{
f02a(x)

f02b(x)

example_data_02(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\description{
\href{https://adventofcode.com/2022/day/2}{Rock Paper Scissors}
}
\examples{
f02a(example_data_02())
f02b(example_data_02())
}
