% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03}
\alias{day03}
\alias{f03a}
\alias{f03b}
\alias{example_data_03}
\title{Day 03: Gear Ratios}
\usage{
f03a(x)

f03b(x)

example_data_03(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f03a(x)} returns .... For Part Two,
\code{f03b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2023/day/3}{Gear Ratios}
}
\details{
\strong{Part One}

You and the Elf eventually reach a \href{https://en.wikipedia.org/wiki/Gondola_lift}{gondola lift}{target="_blank"}
station; he says the gondola lift will take you up to the \emph{water
source}, but this is as far as he can bring you. You go inside.

It doesn\'t take long to find the gondolas, but there seems to be a
problem: they\'re not moving.

\"Aaah!\"

You turn around to see a slightly-greasy Elf with a wrench and a look of
surprise. \"Sorry, I wasn\'t expecting anyone! The gondola lift isn\'t
working right now; it\'ll still be a while before I can fix it.\" You
offer to help.

The engineer explains that an engine part seems to be missing from the
engine, but nobody can figure out which one. If you can \emph{add up all the
part numbers} in the engine schematic, it should be easy to work out
which part is missing.

The engine schematic (your puzzle input) consists of a visual
representation of the engine. There are lots of numbers and symbols you
don\'t really understand, but apparently \emph{any number adjacent to a
symbol}, even diagonally, is a \"part number\" and should be included in
your sum. (Periods (\code{.}) do not count as a symbol.)

Here is an example engine schematic:

\if{html}{\out{<div class="sourceCode">}}\preformatted{467..114..
...*......
..35..633.
......#...
617*......
.....+.58.
..592.....
......755.
...$.*....
.664.598..
}\if{html}{\out{</div>}}

In this schematic, two numbers are \emph{not} part numbers because they are
not adjacent to a symbol: \code{114} (top right) and \code{58} (middle right).
Every other number is adjacent to a symbol and so \emph{is} a part number;
their sum is \emph{\code{4361}}.

Of course, the actual engine schematic is much larger. \emph{What is the sum
of all of the part numbers in the engine schematic?}

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
f03a(example_data_03())
f03b()
}
