% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03}
\alias{day03}
\alias{f03a}
\alias{f03b}
\alias{example_data_03}
\title{Day 03: Lobby}
\usage{
f03a(x)

f03b(x)

example_data_03(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f03a(x)} returns .... For Part Two,
\code{f03b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2025/day/3}{Lobby}
}
\details{
\strong{Part One}

You descend a short staircase, enter the surprisingly vast lobby, and
are quickly cleared by the security checkpoint. When you get to the main
elevators, however, you discover that each one has a red light above it:
they\'re all \emph{offline}.

\"Sorry about that,\" an Elf apologizes as she tinkers with a nearby
control panel. \"Some kind of electrical surge seems to have fried them.
I\'ll try to get them online soon.\"

You explain your need to get further underground. \"Well, you could at
least take the escalator down to the printing department, not that
you\'d get much further than that without the elevators working. That
is, you could if the escalator weren\'t also
[offline]{title="Escalator temporarily stairs. Sorry for the convenience."}.\"

\"But, don\'t worry! It\'s not fried; it just needs power. Maybe you can
get it running while I keep working on the elevators.\"

There are batteries nearby that can supply emergency power to the
escalator for just such an occasion. The batteries are each labeled with
their \href{/2020/day/10}{joltage} rating, a value from \code{1} to \code{9}. You make
a note of their joltage ratings (your puzzle input). For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{987654321111111
811111111111119
234234234234278
818181911112111
}\if{html}{\out{</div>}}

The batteries are arranged into \emph{banks}; each line of digits in your
input corresponds to a single bank of batteries. Within each bank, you
need to turn on \emph{exactly two} batteries; the joltage that the bank
produces is equal to the number formed by the digits on the batteries
you\'ve turned on. For example, if you have a bank like \code{12345} and you
turn on batteries \code{2} and \code{4}, the bank would produce \code{24} jolts. (You
cannot rearrange batteries.)

You\'ll need to find the largest possible joltage each bank can produce.
In the above example:
\itemize{
\item In \emph{\code{98}}\code{7654321111111}, you can make the largest joltage possible,
\emph{\code{98}}, by turning on the first two batteries.
\item In \emph{\code{8}}\code{1111111111111}\emph{\code{9}}, you can make the largest joltage
possible by turning on the batteries labeled \code{8} and \code{9}, producing
\emph{\code{89}} jolts.
\item In \code{2342342342342}\emph{\code{78}}, you can make \emph{\code{78}} by turning on the last
two batteries (marked \code{7} and \code{8}).
\item In \code{818181}\emph{\code{9}}\code{1111}\emph{\code{2}}\code{111}, the largest joltage you can produce
is \emph{\code{92}}.
}

The total output joltage is the sum of the maximum joltage from each
bank, so in this example, the total output joltage is \code{98} + \code{89} +
\code{78} + \code{92} = \emph{\code{357}}.

There are many batteries in front of you. Find the maximum joltage
possible from each bank; \emph{what is the total output joltage?}

\strong{Part Two}

The escalator doesn\'t move. The Elf explains that it probably needs
more joltage to overcome the \href{https://en.wikipedia.org/wiki/Static_friction}{static friction}{target="_blank"}
of the system and hits the big red \"joltage limit safety override\"
button. You lose count of the number of times she needs to confirm
\"yes, I\'m sure\" and decorate the lobby a bit while you wait.

Now, you need to make the largest joltage by turning on \emph{exactly twelve}
batteries within each bank.

The joltage output for the bank is still the number formed by the digits
of the batteries you\'ve turned on; the only difference is that now
there will be \emph{\code{12}} digits in each bank\'s joltage output instead of
two.

Consider again the example from before:

\if{html}{\out{<div class="sourceCode">}}\preformatted{987654321111111
811111111111119
234234234234278
818181911112111
}\if{html}{\out{</div>}}

Now, the joltages are much larger:
\itemize{
\item In \emph{\code{987654321111}}\code{111}, the largest joltage can be found by turning
on everything except some \code{1}s at the end to produce \emph{\code{987654321111}}.
\item In the digit sequence \emph{\code{81111111111}}\code{111}\emph{\code{9}}, the largest joltage
can be found by turning on everything except some \code{1}s, producing
\emph{\code{811111111119}}.
\item In \code{23}\emph{\code{4}}\code{2}\emph{\code{34234234278}}, the largest joltage can be found by
turning on everything except a \code{2} battery, a \code{3} battery, and another
\code{2} battery near the start to produce \emph{\code{434234234278}}.
\item In \emph{\code{8}}\code{1}\emph{\code{8}}\code{1}\emph{\code{8}}\code{1}\emph{\code{911112111}}, the joltage \emph{\code{888911112111}}
is produced by turning on everything except some \code{1}s near the front.
}

The total output joltage is now much larger: \code{987654321111} +
\code{811111111119} + \code{434234234278} + \code{888911112111} = \emph{\code{3121910778619}}.

\emph{What is the new total output joltage?}
}
\examples{
f03a(example_data_03())
f03b()
}
