% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03}
\alias{day03}
\alias{solve03a}
\alias{solve03b}
\alias{example_data_03}
\title{Day 03: Gear Ratios}
\usage{
solve03a(x)

solve03b(x)

example_data_03(example = 1)
}
\arguments{
\item{x}{raw data}

\item{example}{Which example data to use (by position or name)}
}
\description{
\href{https://adventofcode.com/2023/day/3}{Gear Ratios}
}
\examples{
solve03a(example_data_03())
solve03b(example_data_03())
}
