% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{day04}
\alias{day04}
\alias{f04a}
\alias{f04b}
\alias{example_data_04}
\title{Day 04: Printing Department}
\usage{
f04a(x)

f04b(x)

example_data_04(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f04a(x)} returns .... For Part Two,
\code{f04b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2025/day/4}{Printing Department}
}
\details{
\strong{Part One}

You ride the escalator down to the printing department. They\'re clearly
getting ready for Christmas; they have lots of large rolls of paper
everywhere, and there\'s even a massive printer in the corner (to handle
the really [big]{title="This joke is stupid and I love it."} print
jobs).

Decorating here will be easy: they can make their own decorations. What
you really need is a way to get further into the North Pole base while
the elevators are offline.

\"Actually, maybe we can help with that,\" one of the Elves replies when
you ask for help. \"We\'re pretty sure there\'s a cafeteria on the other
side of the back wall. If we could break through the wall, you\'d be
able to keep moving. It\'s too bad all of our forklifts are so busy
moving those big rolls of paper around.\"

If you can optimize the work the forklifts are doing, maybe they would
have time to spare to break through the wall.

The rolls of paper (\code{@}) are arranged on a large grid; the Elves even
have a helpful diagram (your puzzle input) indicating where everything
is located.

For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{..@.@@.
@@.@.@.@
@@@.@.@
@.@@..@.
@.@@.@
.@@@@.@
.@.@.@.@@
@.@@.@@
.@@@@.
@.@.@@.@.
}\if{html}{\out{</div>}}

The forklifts can only access a roll of paper if there are \emph{fewer than
four rolls of paper} in the eight adjacent positions. If you can figure
out which rolls of paper the forklifts can access, they\'ll spend less
time looking and more time breaking down the wall to the cafeteria.

In this example, there are \emph{\code{13}} rolls of paper that can be accessed by
a forklift (marked with \code{x}):

\if{html}{\out{<div class="sourceCode">}}\preformatted{..xx.xx@x.
x@.@.@.@
@@@.x.@
@.@@..@.
x@.@@.@x
.@@@@.@
.@.@.@.@@
x.@@.@@
.@@@@.
x.x.@@.x.
}\if{html}{\out{</div>}}

Consider your complete diagram of the paper roll locations. \emph{How many
rolls of paper can be accessed by a forklift?}

\strong{Part Two}
Now, the Elves just need help accessing as much of the paper as they
can.

Once a roll of paper can be accessed by a forklift, it can be \emph{removed}.
Once a roll of paper is removed, the forklifts might be able to access
\emph{more} rolls of paper, which they might also be able to remove. How many
total rolls of paper could the Elves remove if they keep repeating this
process?

Starting with the same example as above, here is one way you could
remove as many rolls of paper as possible, using highlighted \emph{\code{@}} to
indicate that a roll of paper is about to be removed, and using \code{x} to
indicate that a roll of paper was just removed:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Initial state:
..@.@@.
@@.@.@.@
@@@.@.@
@.@@..@.
@.@@.@
.@@@@.@
.@.@.@.@@
@.@@.@@
.@@@@.
@.@.@@.@.

Remove 13 rolls of paper:
..xx.xx@x.
x@.@.@.@
@@@.x.@
@.@@..@.
x@.@@.@x
.@@@@.@
.@.@.@.@@
x.@@.@@
.@@@@.
x.x.@@.x.

Remove 12 rolls of paper:
.......x..
.@.x.x.@x
x@@...@
x.@@..x.
.@.@@.x.
.x@@@.x
.x.@.@.@@
..@@.@@
.x@@@@.
....@@...

Remove 7 rolls of paper:
..........
.x@.....x.
.@@...xx
..@@....
.x.@@...
..@@@..
...@.@.@x
..@@.@@
..x@@@.
....@@...

Remove 5 rolls of paper:
..........
..x.......
.x@@.....
..@@....
...@@...
..x@@@..
...@.@.@.
..x@.@@x
...@@@.
....@@...

Remove 2 rolls of paper:
..........
..........
..x@.....
..@@....
...@@...
...@@@..
...@.@.@.
...@.@@.
...@@@x.
....@@...

Remove 1 roll of paper:
..........
..........
...@.....
..x@@....
...@@...
...@@@..
...@.@.@.
...@.@@.
...@@@..
....@@...

Remove 1 roll of paper:
..........
..........
...x@.....
...@@....
...@@...
...@@@..
...@.@.@.
...@.@@.
...@@@..
....@@...

Remove 1 roll of paper:
..........
..........
....x.....
...@@....
...@@...
...@@@..
...@.@.@.
...@.@@.
...@@@..
....@@...

Remove 1 roll of paper:
..........
..........
..........
...x@....
...@@...
...@@@..
...@.@.@.
...@.@@.
...@@@..
....@@...
}\if{html}{\out{</div>}}

Stop once no more rolls of paper are accessible by a forklift. In this
example, a total of \emph{\code{43}} rolls of paper can be removed.

Start with your original diagram. \emph{How many rolls of paper in total can
be removed by the Elves and their forklifts?}
}
\examples{
f04a(example_data_04())
f04b()
}
