% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{day04}
\alias{day04}
\alias{f04a}
\alias{f04b}
\alias{example_data_04}
\title{Day 04: Camp Cleanup}
\usage{
f04a(x)

f04b(x)

example_data_04(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\description{
\href{https://adventofcode.com/2022/day/4}{Camp Cleanup}
}
\examples{
f04a(example_data_04())
f04b(example_data_04())
}
