% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{day04}
\alias{day04}
\alias{solve04a}
\alias{solve04b}
\alias{example_data_04}
\title{Day 04: Scratchcards}
\usage{
solve04a(x)

solve04b(x)

example_data_04(example = 1)
}
\arguments{
\item{x}{raw cards}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2023/day/4}{Scratchcards}
}
\examples{
solve04a(example_data_04())
solve04b(example_data_04())
}
