% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05}
\alias{day05}
\alias{f05a}
\alias{f05b}
\alias{example_data_05}
\title{Day 05: Cafeteria}
\usage{
f05a(x)

f05b(x)

example_data_05(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f05a(x)} returns .... For Part Two,
\code{f05b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2025/day/5}{Cafeteria}
}
\details{
\strong{Part One}

As the forklifts break through the wall, the Elves are delighted to
discover that there was a cafeteria on the other side after all.

You can hear a commotion coming from the kitchen. \"At this rate, we
won\'t have any time left to put the wreaths up in the dining hall!\"
Resolute in your quest, you investigate.

\"If only we hadn\'t switched to the new inventory management system
right before Christmas!\" another Elf exclaims. You ask what\'s going
on.

The Elves in the kitchen explain the situation: because of their
complicated new inventory management system, they can\'t figure out
which of their ingredients are \emph{fresh} and which are
[\emph{spoiled}]{title="No, this puzzle does not take place on Gleba. Why do you ask?"}.
When you ask how it works, they give you a copy of their database (your
puzzle input).

The database operates on \emph{ingredient IDs}. It consists of a list of
\emph{fresh ingredient ID ranges}, a blank line, and a list of \emph{available
ingredient IDs}. For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{3-5
10-14
16-20
12-18

1
5
8
11
17
32
}\if{html}{\out{</div>}}

The fresh ID ranges are \emph{inclusive}: the range \code{3-5} means that
ingredient IDs \code{3}, \code{4}, and \code{5} are all \emph{fresh}. The ranges can also
\emph{overlap}; an ingredient ID is fresh if it is in \emph{any} range.

The Elves are trying to determine which of the \emph{available ingredient
IDs} are \emph{fresh}. In this example, this is done as follows:
\itemize{
\item Ingredient ID \code{1} is spoiled because it does not fall into any range.
\item Ingredient ID \code{5} is \emph{fresh} because it falls into range \code{3-5}.
\item Ingredient ID \code{8} is spoiled.
\item Ingredient ID \code{11} is \emph{fresh} because it falls into range \code{10-14}.
\item Ingredient ID \code{17} is \emph{fresh} because it falls into range \code{16-20} as
well as range \code{12-18}.
\item Ingredient ID \code{32} is spoiled.
}

So, in this example, \emph{\code{3}} of the available ingredient IDs are fresh.

Process the database file from the new inventory management system. \emph{How
many of the available ingredient IDs are fresh?}

\strong{Part Two}

The Elves start bringing their spoiled inventory to the trash chute at
the back of the kitchen.

So that they can stop bugging you when they get new inventory, the Elves
would like to know \emph{all} of the IDs that the \emph{fresh ingredient ID
ranges} consider to be \emph{fresh}. An ingredient ID is still considered
fresh if it is in any range.

Now, the second section of the database (the available ingredient IDs)
is irrelevant. Here are the fresh ingredient ID ranges from the above
example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{3-5
10-14
16-20
12-18
}\if{html}{\out{</div>}}

The ingredient IDs that these ranges consider to be fresh are \code{3}, \code{4},
\code{5}, \code{10}, \code{11}, \code{12}, \code{13}, \code{14}, \code{15}, \code{16}, \code{17}, \code{18}, \code{19}, and
\code{20}. So, in this example, the fresh ingredient ID ranges consider a
total of \emph{\code{14}} ingredient IDs to be fresh.

Process the database file again. \emph{How many ingredient IDs are considered
to be fresh according to the fresh ingredient ID ranges?}
}
\examples{
f05a(example_data_05())
f05b()
}
