% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05}
\alias{day05}
\alias{f05a}
\alias{f05b}
\alias{example_data_05}
\title{Day 05: Supply Stacks}
\usage{
f05a(x)

f05b(x)

example_data_05(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\description{
\href{https://adventofcode.com/2022/day/5}{Supply Stacks}
}
\examples{
f05a(example_data_05())
f05b(example_data_05())
}
