% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05}
\alias{day05}
\alias{solve05a}
\alias{solve05b}
\alias{example_data_05}
\title{Day 05: If You Give A Seed A Fertilizer}
\usage{
solve05a(x)

solve05b(x)

example_data_05(example = 1)
}
\arguments{
\item{x}{seed data}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2023/day/5}{If You Give A Seed A Fertilizer}
}
\examples{
solve05a(example_data_05())
solve05b(example_data_05())
}
