% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\name{day06}
\alias{day06}
\alias{f06a}
\alias{f06b}
\alias{example_data_06}
\title{Day 06: Tuning Trouble}
\usage{
f06a(x)

f06b(x)

example_data_06(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f06a(x)} returns .... For Part Two,
\code{f06b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2022/day/6}{Tuning Trouble}
}
\details{
\strong{Part One}

The preparations are finally complete; you and the Elves leave camp on
foot and begin to make your way toward the \emph{star} fruit grove.

As you move through the dense undergrowth, one of the Elves gives you a
handheld \emph{device}. He says that it has many fancy features, but the most
important one to set up right now is the \emph{communication system}.

However, because he\'s heard you have \href{/2016/day/6}{significant}
\href{/2016/day/25}{experience} \href{/2019/day/7}{dealing} \href{/2019/day/9}{with}
\href{/2019/day/16}{signal-based} \href{/2021/day/25}{systems}, he convinced the
other Elves that it would be okay to give you their one malfunctioning
device - surely you\'ll have no problem fixing it.

As if inspired by comedic timing, the device emits a few [colorful
sparks]{title="The magic smoke, on the other hand, seems to be contained... FOR NOW!"}.

To be able to communicate with the Elves, the device needs to \emph{lock on
to their signal}. The signal is a series of seemingly-random characters
that the device receives one at a time.

To fix the communication system, you need to add a subroutine to the
device that detects a \emph{start-of-packet marker} in the datastream. In the
protocol being used by the Elves, the start of a packet is indicated by
a sequence of \emph{four characters that are all different}.

The device will send your subroutine a datastream buffer (your puzzle
input); your subroutine needs to identify the first position where the
four most recently received characters were all different. Specifically,
it needs to report the number of characters from the beginning of the
buffer to the end of the first such four-character marker.

For example, suppose you receive the following datastream buffer:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mjqjpqmgbljsphdztnvjfqwrcgsmlb
}\if{html}{\out{</div>}}

After the first three characters (\code{mjq}) have been received, there
haven\'t been enough characters received yet to find the marker. The
first time a marker could occur is after the fourth character is
received, making the most recent four characters \code{mjqj}. Because \code{j} is
repeated, this isn\'t a marker.

The first time a marker appears is after the \emph{seventh} character
arrives. Once it does, the last four characters received are \code{jpqm},
which are all different. In this case, your subroutine should report the
value \emph{\code{7}}, because the first start-of-packet marker is complete after
7 characters have been processed.

Here are a few more examples:
\itemize{
\item \code{bvwbjplbgvbhsrlpgdmjqwftvncz}: first marker after character \emph{\code{5}}
\item \code{nppdvjthqldpwncqszvftbrmjlhg}: first marker after character \emph{\code{6}}
\item \code{nznrnfrfntjfmvfwmzdfjlvtqnbhcprsg}: first marker after character
\emph{\code{10}}
\item \code{zcfzfwzzqfrljwzlrfnpqdbhtmscgvjw}: first marker after character
\emph{\code{11}}
}

\emph{How many characters need to be processed before the first
start-of-packet marker is detected?}

\strong{Part Two}

Your device\'s communication system is correctly detecting packets, but
still isn\'t working. It looks like it also needs to look for
\emph{messages}.

A \emph{start-of-message marker} is just like a start-of-packet marker,
except it consists of \emph{14 distinct characters} rather than 4.

Here are the first positions of start-of-message markers for all of the
above examples:
\itemize{
\item \code{mjqjpqmgbljsphdztnvjfqwrcgsmlb}: first marker after character
\emph{\code{19}}
\item \code{bvwbjplbgvbhsrlpgdmjqwftvncz}: first marker after character \emph{\code{23}}
\item \code{nppdvjthqldpwncqszvftbrmjlhg}: first marker after character \emph{\code{23}}
\item \code{nznrnfrfntjfmvfwmzdfjlvtqnbhcprsg}: first marker after character
\emph{\code{29}}
\item \code{zcfzfwzzqfrljwzlrfnpqdbhtmscgvjw}: first marker after character
\emph{\code{26}}
}

\emph{How many characters need to be processed before the first
start-of-message marker is detected?}
}
\examples{
f06a(example_data_06())
f06b()
}
