% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\name{day06}
\alias{day06}
\alias{f06a}
\alias{f06b}
\alias{example_data_06}
\title{Day 06: Trash Compactor}
\usage{
f06a(x)

f06b(x)

example_data_06(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f06a(x)} returns .... For Part Two,
\code{f06b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2025/day/6}{Trash Compactor}
}
\details{
\strong{Part One}

After helping the Elves in the kitchen, you were taking a break and
helping them re-enact a movie scene when you over-enthusiastically
jumped into the garbage chute!

A brief fall later, you find yourself in a [garbage
smasher]{title="To your surprise, the smell isn't that bad."}.
Unfortunately, the door\'s been magnetically sealed.

As you try to find a way out, you are approached by a family of
cephalopods! They\'re pretty sure they can get the door open, but it
will take some time. While you wait, they\'re curious if you can help
the youngest cephalopod with her \href{/2021/day/18}{math homework}.

Cephalopod math doesn\'t look that different from normal math. The math
worksheet (your puzzle input) consists of a list of \emph{problems}; each
problem has a group of numbers that need to either be either \emph{added}
(\code{+}) or \emph{multiplied} (\code{*}) together.

However, the problems are arranged a little strangely; they seem to be
presented next to each other in a very long horizontal list. For
example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{123 328  51 64
 45 64  387 23
  6 98  215 314
*   +   *   +
}\if{html}{\out{</div>}}

Each problem\'s numbers are arranged vertically; at the bottom of the
problem is the symbol for the operation that needs to be performed.
Problems are separated by a full column of only spaces. The left/right
alignment of numbers within each problem can be ignored.

So, this worksheet contains four problems:
\itemize{
\item \code{123} \* \code{45} \* \code{6} = \emph{\code{33210}}
\item \code{328} + \code{64} + \code{98} = \emph{\code{490}}
\item \code{51} \* \code{387} \* \code{215} = \emph{\code{4243455}}
\item \code{64} + \code{23} + \code{314} = \emph{\code{401}}
}

To check their work, cephalopod students are given the \emph{grand total} of
adding together all of the answers to the individual problems. In this
worksheet, the grand total is \code{33210} + \code{490} + \code{4243455} + \code{401} =
\emph{\code{4277556}}.

Of course, the actual worksheet is \emph{much} wider. You\'ll need to make
sure to unroll it completely so that you can read the problems clearly.

Solve the problems on the math worksheet. \emph{What is the grand total found
by adding together all of the answers to the individual problems?}

\strong{Part Two}

The big cephalopods come back to check on how things are going. When
they see that your grand total doesn\'t match the one expected by the
worksheet, they realize they forgot to explain how to read cephalopod
math.

Cephalopod math is written \emph{right-to-left in columns}. Each number is
given in its own column, with the most significant digit at the top and
the least significant digit at the bottom. (Problems are still separated
with a column consisting only of spaces, and the symbol at the bottom of
the problem is still the operator to use.)

Here\'s the example worksheet again:

\if{html}{\out{<div class="sourceCode">}}\preformatted{123 328  51 64
 45 64  387 23
  6 98  215 314
*   +   *   +
}\if{html}{\out{</div>}}

Reading the problems right-to-left one column at a time, the problems
are now quite different:
\itemize{
\item The rightmost problem is \code{4} + \code{431} + \code{623} = \emph{\code{1058}}
\item The second problem from the right is \code{175} \* \code{581} \* \code{32} =
\emph{\code{3253600}}
\item The third problem from the right is \code{8} + \code{248} + \code{369} = \emph{\code{625}}
\item Finally, the leftmost problem is \code{356} \* \code{24} \* \code{1} = \emph{\code{8544}}
}

Now, the grand total is \code{1058} + \code{3253600} + \code{625} + \code{8544} =
\emph{\code{3263827}}.

Solve the problems on the math worksheet again. \emph{What is the grand total
found by adding together all of the answers to the individual problems?}
}
\examples{
f06a(example_data_06())
f06b()
}
