% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\name{day06}
\alias{day06}
\alias{f06a}
\alias{f06b}
\alias{example_data_06}
\title{Day 06: Tuning Trouble}
\usage{
f06a(x)

f06b(x)

example_data_06(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\description{
\href{https://adventofcode.com/2022/day/6}{Tuning Trouble}
}
\examples{
f06a(example_data_06())
f06b(example_data_06())
}
