% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\name{day06}
\alias{day06}
\alias{solve06a}
\alias{solve06b}
\alias{example_data_06}
\title{Day 06: Wait For It}
\usage{
solve06a(x)

solve06b(x)

example_data_06(example = 1)
}
\arguments{
\item{x}{raw races}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2023/day/6}{Wait For It}
}
\examples{
solve06a(example_data_06())
solve06b(example_data_06())
}
