% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day07.R
\name{day07}
\alias{day07}
\alias{f07a}
\alias{f07b}
\alias{example_data_07}
\title{Day 07: Camel Cards}
\usage{
f07a(x)

f07b(x)

example_data_07(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f07a(x)} returns .... For Part Two,
\code{f07b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2023/day/7}{Camel Cards}
}
\details{
\strong{Part One}

Your all-expenses-paid trip turns out to be a one-way, five-minute ride
in an \href{https://en.wikipedia.org/wiki/Airship}{airship}{target="_blank"}.
(At least it\'s a [\emph{cool}
airship]{title="Please only read this sentence while listening to 'The Airship Blackjack' from the Final Fantasy 6 soundtrack."}!)
It drops you off at the edge of a vast desert and descends back to
Island Island.

\"Did you bring the parts?\"

You turn around to see an Elf completely covered in white clothing,
wearing goggles, and riding a large
\href{https://en.wikipedia.org/wiki/Dromedary}{camel}{target="_blank"}.

\"Did you bring the parts?\" she asks again, louder this time. You
aren\'t sure what parts she\'s looking for; you\'re here to figure out
why the sand stopped.

\"The parts! For the sand, yes! Come with me; I will show you.\" She
beckons you onto the camel.

After riding a bit across the sands of Desert Island, you can see what
look like very large rocks covering half of the horizon. The Elf
explains that the rocks are all along the part of Desert Island that is
directly above Island Island, making it hard to even get there.
Normally, they use big machines to move the rocks and filter the sand,
but the machines have broken down because Desert Island recently stopped
receiving the \emph{parts} they need to fix the machines.

You\'ve already assumed it\'ll be your job to figure out why the parts
stopped when she asks if you can help. You agree automatically.

Because the journey will take a few days, she offers to teach you the
game of \emph{Camel Cards}. Camel Cards is sort of similar to
\href{https://en.wikipedia.org/wiki/List_of_poker_hands}{poker}{target="_blank"}
except it\'s designed to be easier to play while riding a camel.

In Camel Cards, you get a list of \emph{hands}, and your goal is to order
them based on the \emph{strength} of each hand. A hand consists of \emph{five
cards} labeled one of \code{A}, \code{K}, \code{Q}, \code{J}, \code{T}, \code{9}, \code{8}, \code{7}, \code{6}, \code{5},
\code{4}, \code{3}, or \code{2}. The relative strength of each card follows this order,
where \code{A} is the highest and \code{2} is the lowest.

Every hand is exactly one \emph{type}. From strongest to weakest, they are:
\itemize{
\item \emph{Five of a kind}, where all five cards have the same label: \code{AAAAA}
\item \emph{Four of a kind}, where four cards have the same label and one card
has a different label: \code{AA8AA}
\item \emph{Full house}, where three cards have the same label, and the
remaining two cards share a different label: \code{23332}
\item \emph{Three of a kind}, where three cards have the same label, and the
remaining two cards are each different from any other card in the
hand: \code{TTT98}
\item \emph{Two pair}, where two cards share one label, two other cards share a
second label, and the remaining card has a third label: \code{23432}
\item \emph{One pair}, where two cards share one label, and the other three
cards have a different label from the pair and each other: \code{A23A4}
\item \emph{High card}, where all cards\' labels are distinct: \code{23456}
}

Hands are primarily ordered based on type; for example, every \emph{full
house} is stronger than any \emph{three of a kind}.

If two hands have the same type, a second ordering rule takes effect.
Start by comparing the \emph{first card in each hand}. If these cards are
different, the hand with the stronger first card is considered stronger.
If the first card in each hand have the \emph{same label}, however, then move
on to considering the \emph{second card in each hand}. If they differ, the
hand with the higher second card wins; otherwise, continue with the
third card in each hand, then the fourth, then the fifth.

So, \code{33332} and \verb{2AAAA} are both \emph{four of a kind} hands, but \code{33332} is
stronger because its first card is stronger. Similarly, \code{77888} and
\code{77788} are both a \emph{full house}, but \code{77888} is stronger because its
third card is stronger (and both hands have the same first and second
card).

To play Camel Cards, you are given a list of hands and their
corresponding \emph{bid} (your puzzle input). For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{32T3K 765
T55J5 684
KK677 28
KTJJT 220
QQQJA 483
}\if{html}{\out{</div>}}

This example shows five hands; each hand is followed by its \emph{bid}
amount. Each hand wins an amount equal to its bid multiplied by its
\emph{rank}, where the weakest hand gets rank 1, the second-weakest hand gets
rank 2, and so on up to the strongest hand. Because there are five hands
in this example, the strongest hand will have rank 5 and its bid will be
multiplied by 5.

So, the first step is to put the hands in order of strength:
\itemize{
\item \verb{32T3K} is the only \emph{one pair} and the other hands are all a
stronger type, so it gets rank \emph{1}.
\item \code{KK677} and \code{KTJJT} are both \emph{two pair}. Their first cards both have
the same label, but the second card of \code{KK677} is stronger (\code{K} vs
\code{T}), so \code{KTJJT} gets rank \emph{2} and \code{KK677} gets rank \emph{3}.
\item \code{T55J5} and \code{QQQJA} are both \emph{three of a kind}. \code{QQQJA} has a
stronger first card, so it gets rank \emph{5} and \code{T55J5} gets rank \emph{4}.
}

Now, you can determine the total winnings of this set of hands by adding
up the result of multiplying each hand\'s bid with its rank (\code{765} \*
1 + \code{220} \* 2 + \code{28} \* 3 + \code{684} \* 4 + \code{483} \* 5). So the \emph{total
winnings} in this example are \emph{\code{6440}}.

Find the rank of every hand in your set. \emph{What are the total winnings?}

\strong{Part Two}

To make things a little more interesting, the Elf introduces one
additional rule. Now, \code{J} cards are
\href{https://en.wikipedia.org/wiki/Joker_(playing_card)}{jokers}{target="_blank"} -
wildcards that can act like whatever card would make the hand the
strongest type possible.

To balance this, \emph{\code{J} cards are now the weakest} individual cards,
weaker even than \code{2}. The other cards stay in the same order: \code{A}, \code{K},
\code{Q}, \code{T}, \code{9}, \code{8}, \code{7}, \code{6}, \code{5}, \code{4}, \code{3}, \code{2}, \code{J}.

\code{J} cards can pretend to be whatever card is best for the purpose of
determining hand type; for example, \code{QJJQ2} is now considered \emph{four of a
kind}. However, for the purpose of breaking ties between two hands of
the same type, \code{J} is always treated as \code{J}, not the card it\'s
pretending to be: \code{JKKK2} is weaker than \code{QQQQ2} because \code{J} is weaker
than \code{Q}.

Now, the above example goes very differently:

\if{html}{\out{<div class="sourceCode">}}\preformatted{32T3K 765
T55J5 684
KK677 28
KTJJT 220
QQQJA 483
}\if{html}{\out{</div>}}
\itemize{
\item \verb{32T3K} is still the only \emph{one pair}; it doesn\'t contain any
jokers, so its strength doesn\'t increase.
\item \code{KK677} is now the only \emph{two pair}, making it the second-weakest
hand.
\item \code{T55J5}, \code{KTJJT}, and \code{QQQJA} are now all \emph{four of a kind}! \code{T55J5}
gets rank 3, \code{QQQJA} gets rank 4, and \code{KTJJT} gets rank 5.
}

With the new joker rule, the total winnings in this example are
\emph{\code{5905}}.

Using the new joker rule, find the rank of every hand in your set. \emph{What
are the new total winnings?}
}
\examples{
f07a(example_data_07())
f07b()
}
