% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day07.R
\name{day07}
\alias{day07}
\alias{solve07a}
\alias{solve07b}
\alias{example_data_07}
\title{Day 07: The Treachery of Whales}
\usage{
solve07a(x)

solve07b(x)

example_data_07(example = 1)
}
\arguments{
\item{x}{raw crab position}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2021/day/7}{The Treachery of Whales}
}
\examples{
solve07a(example_data_07())
solve07b(example_data_07())
}
