% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day08.R
\name{day08}
\alias{day08}
\alias{solve08a}
\alias{solve08b}
\alias{example_data_08}
\title{Day 08: Seven Segment Search}
\usage{
solve08a(x)

solve08b(x)

example_data_08(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2021/day/8}{Seven Segment Search}
}
\examples{
solve08a(example_data_08())
solve08b(example_data_08())
}
