% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day08.R
\name{day08}
\alias{day08}
\alias{f08a}
\alias{f08b}
\alias{example_data_08}
\title{Day 08: Treetop Tree House}
\usage{
f08a(x)

f08b(x)

example_data_08(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\description{
\href{https://adventofcode.com/2022/day/8}{Treetop Tree House}
}
\examples{
f08a(example_data_08())
f08b(example_data_08())
}
