% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day09.R
\name{day09}
\alias{day09}
\alias{f09a}
\alias{f09b}
\alias{example_data_09}
\title{Day 09: Rope Bridge}
\usage{
f09a(x)

f09b(x)

example_data_09(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f09a(x)} returns .... For Part Two,
\code{f09b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2022/day/9}{Rope Bridge}
}
\details{
\strong{Part One}

This rope bridge creaks as you walk along it. You aren\'t sure how old
it is, or whether it can even support your weight.

It seems to support the Elves just fine, though. The bridge spans a
gorge which was carved out by the massive river far below you.

You step carefully; as you do, the ropes stretch and twist. You decide
to distract yourself by modeling rope physics; maybe you can even figure
out where \emph{not} to step.

Consider a rope with a knot at each end; these knots mark the \emph{head} and
the \emph{tail} of the rope. If the head moves far enough away from the tail,
the tail is pulled toward the head.

Due to nebulous reasoning involving \href{https://en.wikipedia.org/wiki/Planck_units#Planck_length}{Planck lengths}{target="_blank"},
you should be able to model the positions of the knots on a
two-dimensional grid. Then, by following a hypothetical \emph{series of
motions} (your puzzle input) for the head, you can determine how the
tail will move.

[Due to the aforementioned Planck
lengths]{title="I'm an engineer, not a physicist!"}, the rope must be
quite short; in fact, the head (\code{H}) and tail (\code{T}) must \emph{always be
touching} (diagonally adjacent and even overlapping both count as
touching):

\if{html}{\out{<div class="sourceCode">}}\preformatted{....
.TH.
....

....
.H..
..T.
....

...
.H. (H covers T)
...
}\if{html}{\out{</div>}}

If the head is ever two steps directly up, down, left, or right from the
tail, the tail must also move one step in that direction so it remains
close enough:

\if{html}{\out{<div class="sourceCode">}}\preformatted{.....    .....    .....
.TH.. -> .T.H. -> ..TH.
.....    .....    .....

...    ...    ...
.T.    .T.    ...
.H. -> ... -> .T.
...    .H.    .H.
...    ...    ...
}\if{html}{\out{</div>}}

Otherwise, if the head and tail aren\'t touching and aren\'t in the same
row or column, the tail always moves one step diagonally to keep up:

\if{html}{\out{<div class="sourceCode">}}\preformatted{.....    .....    .....
.....    ..H..    ..H..
..H.. -> ..... -> ..T..
.T...    .T...    .....
.....    .....    .....

.....    .....    .....
.....    .....    .....
..H.. -> ...H. -> ..TH.
.T...    .T...    .....
.....    .....    .....
}\if{html}{\out{</div>}}

You just need to work out where the tail goes as the head follows a
series of motions. Assume the head and the tail both start at the same
position, overlapping.

For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{R 4
U 4
L 3
D 1
R 4
D 1
L 5
R 2
}\if{html}{\out{</div>}}

This series of motions moves the head \emph{right} four steps, then \emph{up} four
steps, then \emph{left} three steps, then \emph{down} one step, and so on. After
each step, you\'ll need to update the position of the tail if the step
means the head is no longer adjacent to the tail. Visually, these
motions occur as follows (\code{s} marks the starting position as a reference
point):

\if{html}{\out{<div class="sourceCode">}}\preformatted{== Initial State ==

......
......
......
......
H.....  (H covers T, s)

== R 4 ==

......
......
......
......
TH....  (T covers s)

......
......
......
......
sTH...

......
......
......
......
s.TH..

......
......
......
......
s..TH.

== U 4 ==

......
......
......
....H.
s..T..

......
......
....H.
....T.
s.....

......
....H.
....T.
......
s.....

....H.
....T.
......
......
s.....

== L 3 ==

...H..
....T.
......
......
s.....

..HT..
......
......
......
s.....

.HT...
......
......
......
s.....

== D 1 ==

..T...
.H....
......
......
s.....

== R 4 ==

..T...
..H...
......
......
s.....

..T...
...H..
......
......
s.....

......
...TH.
......
......
s.....

......
....TH
......
......
s.....

== D 1 ==

......
....T.
.....H
......
s.....

== L 5 ==

......
....T.
....H.
......
s.....

......
....T.
...H..
......
s.....

......
......
..HT..
......
s.....

......
......
.HT...
......
s.....

......
......
HT....
......
s.....

== R 2 ==

......
......
.H....  (H covers T)
......
s.....

......
......
.TH...
......
s.....
}\if{html}{\out{</div>}}

After simulating the rope, you can count up all of the positions the
\emph{tail visited at least once}. In this diagram, \code{s} again marks the
starting position (which the tail also visited) and \verb{#} marks other
positions the tail visited:

\if{html}{\out{<div class="sourceCode">}}\preformatted{..##..
...##.
.####.
....#.
s###..
}\if{html}{\out{</div>}}

So, there are \emph{\code{13}} positions the tail visited at least once.

Simulate your complete hypothetical series of motions. \emph{How many
positions does the tail of the rope visit at least once?}

\strong{Part Two}

A rope snaps! Suddenly, the river is getting a lot closer than you
remember. The bridge is still there, but some of the ropes that broke
are now whipping toward you as you fall through the air!

The ropes are moving too quickly to grab; you only have a few seconds to
choose how to arch your body to avoid being hit. Fortunately, your
simulation can be extended to support longer ropes.

Rather than two knots, you now must simulate a rope consisting of \emph{ten}
knots. One knot is still the head of the rope and moves according to the
series of motions. Each knot further down the rope follows the knot in
front of it using the same rules as before.

Using the same series of motions as the above example, but with the
knots marked \code{H}, \code{1}, \code{2}, \..., \code{9}, the motions now occur as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{== Initial State ==

......
......
......
......
H.....  (H covers 1, 2, 3, 4, 5, 6, 7, 8, 9, s)

== R 4 ==

......
......
......
......
1H....  (1 covers 2, 3, 4, 5, 6, 7, 8, 9, s)

......
......
......
......
21H...  (2 covers 3, 4, 5, 6, 7, 8, 9, s)

......
......
......
......
321H..  (3 covers 4, 5, 6, 7, 8, 9, s)

......
......
......
......
4321H.  (4 covers 5, 6, 7, 8, 9, s)

== U 4 ==

......
......
......
....H.
4321..  (4 covers 5, 6, 7, 8, 9, s)

......
......
....H.
.4321.
5.....  (5 covers 6, 7, 8, 9, s)

......
....H.
....1.
.432..
5.....  (5 covers 6, 7, 8, 9, s)

....H.
....1.
..432.
.5....
6.....  (6 covers 7, 8, 9, s)

== L 3 ==

...H..
....1.
..432.
.5....
6.....  (6 covers 7, 8, 9, s)

..H1..
...2..
..43..
.5....
6.....  (6 covers 7, 8, 9, s)

.H1...
...2..
..43..
.5....
6.....  (6 covers 7, 8, 9, s)

== D 1 ==

..1...
.H.2..
..43..
.5....
6.....  (6 covers 7, 8, 9, s)

== R 4 ==

..1...
..H2..
..43..
.5....
6.....  (6 covers 7, 8, 9, s)

..1...
...H..  (H covers 2)
..43..
.5....
6.....  (6 covers 7, 8, 9, s)

......
...1H.  (1 covers 2)
..43..
.5....
6.....  (6 covers 7, 8, 9, s)

......
...21H
..43..
.5....
6.....  (6 covers 7, 8, 9, s)

== D 1 ==

......
...21.
..43.H
.5....
6.....  (6 covers 7, 8, 9, s)

== L 5 ==

......
...21.
..43H.
.5....
6.....  (6 covers 7, 8, 9, s)

......
...21.
..4H..  (H covers 3)
.5....
6.....  (6 covers 7, 8, 9, s)

......
...2..
..H1..  (H covers 4; 1 covers 3)
.5....
6.....  (6 covers 7, 8, 9, s)

......
...2..
.H13..  (1 covers 4)
.5....
6.....  (6 covers 7, 8, 9, s)

......
......
H123..  (2 covers 4)
.5....
6.....  (6 covers 7, 8, 9, s)

== R 2 ==

......
......
.H23..  (H covers 1; 2 covers 4)
.5....
6.....  (6 covers 7, 8, 9, s)

......
......
.1H3..  (H covers 2, 4)
.5....
6.....  (6 covers 7, 8, 9, s)
}\if{html}{\out{</div>}}

Now, you need to keep track of the positions the new tail, \code{9}, visits.
In this example, the tail never moves, and so it only visits \emph{\code{1}}
position. However, \emph{be careful}: more types of motion are possible than
before, so you might want to visually compare your simulated rope to the
one above.

Here\'s a larger example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{R 5
U 8
L 8
D 3
R 17
D 10
L 25
U 20
}\if{html}{\out{</div>}}

These motions occur as follows (individual steps are not shown):

\if{html}{\out{<div class="sourceCode">}}\preformatted{== Initial State ==

..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
...........H..............  (H covers 1, 2, 3, 4, 5, 6, 7, 8, 9, s)
..........................
..........................
..........................
..........................
..........................

== R 5 ==

..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
...........54321H.........  (5 covers 6, 7, 8, 9, s)
..........................
..........................
..........................
..........................
..........................

== U 8 ==

..........................
..........................
..........................
..........................
..........................
..........................
..........................
................H.........
................1.........
................2.........
................3.........
...............54.........
..............6...........
.............7............
............8.............
...........9..............  (9 covers s)
..........................
..........................
..........................
..........................
..........................

== L 8 ==

..........................
..........................
..........................
..........................
..........................
..........................
..........................
........H1234.............
............5.............
............6.............
............7.............
............8.............
............9.............
..........................
..........................
...........s..............
..........................
..........................
..........................
..........................
..........................

== D 3 ==

..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
.........2345.............
........1...6.............
........H...7.............
............8.............
............9.............
..........................
..........................
...........s..............
..........................
..........................
..........................
..........................
..........................

== R 17 ==

..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
................987654321H
..........................
..........................
..........................
..........................
...........s..............
..........................
..........................
..........................
..........................
..........................

== D 10 ==

..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
...........s.........98765
.........................4
.........................3
.........................2
.........................1
.........................H

== L 25 ==

..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
...........s..............
..........................
..........................
..........................
..........................
H123456789................

== U 20 ==

H.........................
1.........................
2.........................
3.........................
4.........................
5.........................
6.........................
7.........................
8.........................
9.........................
..........................
..........................
..........................
..........................
..........................
...........s..............
..........................
..........................
..........................
..........................
..........................
}\if{html}{\out{</div>}}

Now, the tail (\code{9}) visits \emph{\code{36}} positions (including \code{s}) at least
once:

\if{html}{\out{<div class="sourceCode">}}\preformatted{..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
..........................
#.........................
#.............###.........
#............#...#........
.#..........#.....#.......
..#..........#.....#......
...#........#.......#.....
....#......s.........#....
.....#..............#.....
......#............#......
.......#..........#.......
........#........#........
.........########.........
}\if{html}{\out{</div>}}

Simulate your complete series of motions on a larger rope with ten
knots. \emph{How many positions does the tail of the rope visit at least
once?}
}
\examples{
f09a(example_data_09())
f09b()
}
