% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day09.R
\name{day09}
\alias{day09}
\alias{f09a}
\alias{f09b}
\alias{example_data_09}
\title{Day 09: Rope Bridge}
\usage{
f09a(x)

f09b(x)

example_data_09(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\description{
\href{https://adventofcode.com/2022/day/9}{Rope Bridge}
}
\examples{
f09a(example_data_09())
f09b(example_data_09())
}
