% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day09.R
\name{day09}
\alias{day09}
\alias{solve09a}
\alias{solve09b}
\alias{example_data_09}
\title{Day 09: Mirage Maintenance}
\usage{
solve09a(x)

solve09b(x)

example_data_09(example = 1)
}
\arguments{
\item{x}{raw OASIS report}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2023/day/9}{Mirage Maintenance}
}
\examples{
solve09a(example_data_09())
solve09b(example_data_09())
}
