% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{day10}
\alias{day10}
\alias{solve10a}
\alias{solve10b}
\alias{example_data_10}
\title{Day 10: Pipe Maze}
\usage{
solve10a(x)

solve10b(x)

example_data_10(example = 1)
}
\arguments{
\item{x}{raw pipes}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2023/day/10}{Pipe Maze}
}
\examples{
solve10a(example_data_10())
solve10b(example_data_10())
}
