% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day11.R
\name{day11}
\alias{day11}
\alias{solve11a}
\alias{solve11b}
\alias{example_data_11}
\title{Day 11: Cosmic Expansion}
\usage{
solve11a(x)

solve11b(x)

example_data_11(example = 1)
}
\arguments{
\item{x}{raw universe}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2023/day/11}{Cosmic Expansion}
}
\examples{
solve11a(example_data_11())
solve11b(example_data_11())
}
