% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day12.R
\name{day12}
\alias{day12}
\alias{f12a}
\alias{f12b}
\alias{example_data_12}
\title{Day 12: Hill Climbing Algorithm}
\usage{
f12a(x)

f12b(x)

example_data_12(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f12a(x)} returns .... For Part Two,
\code{f12b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2022/day/12}{Hill Climbing Algorithm}
}
\details{
\strong{Part One}

You try contacting the Elves using your [handheld
device]{title="When you look up the specs for your handheld device, every field just says \"plot\"."},
but the river you\'re following must be too low to get a decent signal.

You ask the device for a heightmap of the surrounding area (your puzzle
input). The heightmap shows the local area from above broken into a
grid; the elevation of each square of the grid is given by a single
lowercase letter, where \code{a} is the lowest elevation, \code{b} is the
next-lowest, and so on up to the highest elevation, \code{z}.

Also included on the heightmap are marks for your current position (\code{S})
and the location that should get the best signal (\code{E}). Your current
position (\code{S}) has elevation \code{a}, and the location that should get the
best signal (\code{E}) has elevation \code{z}.

You\'d like to reach \code{E}, but to save energy, you should do it in \emph{as
few steps as possible}. During each step, you can move exactly one
square up, down, left, or right. To avoid needing to get out your
climbing gear, the elevation of the destination square can be \emph{at most
one higher} than the elevation of your current square; that is, if your
current elevation is \code{m}, you could step to elevation \code{n}, but not to
elevation \code{o}. (This also means that the elevation of the destination
square can be much lower than the elevation of your current square.)

For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Sabqponm
abcryxxl
accszExk
acctuvwj
abdefghi
}\if{html}{\out{</div>}}

Here, you start in the top-left corner; your goal is near the middle.
You could start by moving down or right, but eventually you\'ll need to
head toward the \code{e} at the bottom. From there, you can spiral around to
the goal:

\if{html}{\out{<div class="sourceCode">}}\preformatted{v..v<<<<
>v.vv<<^
.>vv>E^^
..v>>>^^
..>>>>>^
}\if{html}{\out{</div>}}

In the above diagram, the symbols indicate whether the path exits each
square moving up (\code{^}), down (\code{v}), left (\code{<}), or right (\code{>}). The
location that should get the best signal is still \code{E}, and \code{.} marks
unvisited squares.

This path reaches the goal in \emph{\code{31}} steps, the fewest possible.

\emph{What is the fewest steps required to move from your current position to
the location that should get the best signal?}

\strong{Part Two}

As you walk up the hill, you suspect that the Elves will want to turn
this into a hiking trail. The beginning isn\'t very scenic, though;
perhaps you can find a better starting point.

To maximize exercise while hiking, the trail should start as low as
possible: elevation \code{a}. The goal is still the square marked \code{E}.
However, the trail should still be direct, taking the fewest steps to
reach its goal. So, you\'ll need to find the shortest path from \emph{any
square at elevation \code{a}} to the square marked \code{E}.

Again consider the example from above:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Sabqponm
abcryxxl
accszExk
acctuvwj
abdefghi
}\if{html}{\out{</div>}}

Now, there are six choices for starting position (five marked \code{a}, plus
the square marked \code{S} that counts as being at elevation \code{a}). If you
start at the bottom-left square, you can reach the goal most quickly:

\if{html}{\out{<div class="sourceCode">}}\preformatted{...v<<<<
...vv<<^
...v>E^^
.>v>>>^^
>^>>>>>^
}\if{html}{\out{</div>}}

This path reaches the goal in only \emph{\code{29}} steps, the fewest possible.

\emph{What is the fewest steps required to move starting from any square with
elevation \code{a} to the location that should get the best signal?}
}
\examples{
f12a(example_data_12())
f12b()
}
