% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day12.R
\name{day12}
\alias{day12}
\alias{solve12a}
\alias{solve12b}
\alias{example_data_12}
\title{Day 12: Hill Climbing Algorithm}
\usage{
solve12a(x)

solve12b(x)

example_data_12(example = 1)
}
\arguments{
\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2022/day/12}{Hill Climbing Algorithm}
}
\examples{
solve12a(example_data_12())
solve12b(example_data_12())
}
