% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day14.R
\name{day14}
\alias{day14}
\alias{solve14a}
\alias{solve14b}
\alias{example_data_14}
\title{Day 14: Parabolic Reflector Dish}
\usage{
solve14a(x)

solve14b(x)

example_data_14(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2023/day/14}{Parabolic Reflector Dish}
}
\examples{
solve14a(example_data_14())
solve14b(example_data_14())
}
