% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day15.R
\name{day15}
\alias{day15}
\alias{solve15a}
\alias{solve15b}
\alias{example_data_15}
\title{Day 15: Beacon Exclusion Zone}
\usage{
solve15a(x, y0 = 2000000)

solve15b(x)

example_data_15(example = 1)
}
\arguments{
\item{x}{some data}

\item{y0}{row where to count (10 for example data, 2000000 for real input).}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2022/day/15}{Beacon Exclusion Zone}
}
