% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day15.R
\name{day15}
\alias{day15}
\alias{solve15a}
\alias{solve15b}
\alias{example_data_15}
\title{Day 15: Lens Library}
\usage{
solve15a(x)

solve15b(x)

example_data_15(example = 1)
}
\arguments{
\item{x}{raw instructions}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2023/day/15}{Lens Library}
}
\examples{
solve15a(example_data_15())
solve15b(example_data_15())
}
