% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day17.R
\name{day17}
\alias{day17}
\alias{solve17a}
\alias{solve17b}
\alias{example_data_17}
\title{Day 17: Pyroclastic Flow}
\usage{
solve17a(x)

solve17b(x)

example_data_17(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2022/day/17}{Pyroclastic Flow}
}
\examples{
solve17a(example_data_17())
solve17b(example_data_17())
}
