% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day18.R
\name{day18}
\alias{day18}
\alias{f18a}
\alias{f18b}
\alias{example_data_18}
\title{Day 18: Lavaduct Lagoon}
\usage{
f18a(x, size = 1000)

f18b(x)

example_data_18(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f18a(x)} returns .... For Part Two,
\code{f18b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2023/day/18}{Lavaduct Lagoon}
}
\details{
\strong{Part One}

Thanks to your efforts, the machine parts factory is one of the first
factories up and running since the lavafall came back. However, to catch
up with the large backlog of parts requests, the factory will also need
a \emph{large supply of lava} for a while; the Elves have already started
creating a large lagoon nearby for this purpose.

However, they aren\'t sure the lagoon will be big enough; they\'ve asked
you to take a look at the \emph{dig plan} (your puzzle input). For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{R 6 (#70c710)
D 5 (#0dc571)
L 2 (#5713f0)
D 2 (#d2c081)
R 2 (#59c680)
D 2 (#411b91)
L 5 (#8ceee2)
U 2 (#caa173)
L 1 (#1b58a2)
U 2 (#caa171)
R 2 (#7807d2)
U 3 (#a77fa3)
L 2 (#015232)
U 2 (#7a21e3)
}\if{html}{\out{</div>}}

The digger starts in a 1 meter cube hole in the ground. They then dig
the specified number of meters \emph{up} (\code{U}), \emph{down} (\code{D}), \emph{left} (\code{L}),
or \emph{right} (\code{R}), clearing full 1 meter cubes as they go. The directions
are given as seen from above, so if \"up\" were north, then \"right\"
would be east, and so on. Each trench is also listed with \emph{the color
that the edge of the trench should be painted} as an \href{https://en.wikipedia.org/wiki/RGB_color_model#Numeric_representations}{RGB hexadecimal color code}{target="_blank"}.

When viewed from above, the above example dig plan would result in the
following loop of \emph{trench} (\verb{#}) having been dug out from otherwise
\emph{ground-level terrain} (\code{.}):

\if{html}{\out{<div class="sourceCode">}}\preformatted{#######
#.....#
###...#
..#...#
..#...#
###.###
#...#..
##..###
.#....#
.######
}\if{html}{\out{</div>}}

At this point, the trench could contain 38 cubic meters of lava.
However, this is just the edge of the lagoon; the next step is to \emph{dig
out the interior} so that it is one meter deep as well:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#######
#######
#######
..#####
..#####
#######
#####..
#######
.######
.######
}\if{html}{\out{</div>}}

Now, the lagoon can contain a much more respectable \emph{\code{62}} cubic meters
of lava. While the interior is dug out, the edges are also painted
according to the color codes in the dig plan.

The Elves are concerned the lagoon won\'t be large enough; if they
follow their dig plan, \emph{how many cubic meters of lava could it hold?}

\strong{Part Two}

The Elves were right to be concerned; the planned lagoon would be \emph{much
too small}.

After a few minutes, someone realizes what happened; someone
\emph{[swapped]{title="Futuristic sprintf()?"} the color and instruction
parameters} when producing the dig plan. They don\'t have time to fix
the bug; one of them asks if you can \emph{extract the correct instructions}
from the hexadecimal codes.

Each hexadecimal code is \emph{six hexadecimal digits} long. The first five
hexadecimal digits encode the \emph{distance in meters} as a five-digit
hexadecimal number. The last hexadecimal digit encodes the \emph{direction to
dig}: \code{0} means \code{R}, \code{1} means \code{D}, \code{2} means \code{L}, and \code{3} means \code{U}.

So, in the above example, the hexadecimal codes can be converted into
the true instructions:
\itemize{
\item \verb{#70c710} = \verb{R 461937}
\item \verb{#0dc571} = \verb{D 56407}
\item \verb{#5713f0} = \verb{R 356671}
\item \verb{#d2c081} = \verb{D 863240}
\item \verb{#59c680} = \verb{R 367720}
\item \verb{#411b91} = \verb{D 266681}
\item \verb{#8ceee2} = \verb{L 577262}
\item \verb{#caa173} = \verb{U 829975}
\item \verb{#1b58a2} = \verb{L 112010}
\item \verb{#caa171} = \verb{D 829975}
\item \verb{#7807d2} = \verb{L 491645}
\item \verb{#a77fa3} = \verb{U 686074}
\item \verb{#015232} = \verb{L 5411}
\item \verb{#7a21e3} = \verb{U 500254}
}

Digging out this loop and its interior produces a lagoon that can hold
an impressive \emph{\code{952408144115}} cubic meters of lava.

Convert the hexadecimal color codes into the correct instructions; if
the Elves follow this new dig plan, \emph{how many cubic meters of lava could
the lagoon hold?}
}
\examples{
f18a(example_data_18())
f18b()
}
