% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day18.R
\name{day18}
\alias{day18}
\alias{solve18a}
\alias{solve18b}
\alias{example_data_18}
\title{Day 18: Boiling Boulders}
\usage{
solve18a(x)

solve18b(x)

example_data_18(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\description{
\href{https://adventofcode.com/2022/day/18}{Boiling Boulders}
}
\examples{
solve18a(example_data_18())
solve18b(example_data_18())
}
