% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day18.R
\name{day18}
\alias{day18}
\alias{solve18a}
\alias{example_data_18}
\title{Day 18: Lavaduct Lagoon}
\usage{
solve18a(x)

example_data_18(example = 1)
}
\arguments{
\item{x}{raw dig plan}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2023/day/18}{Lavaduct Lagoon}
}
\examples{
solve18a(example_data_18())
solve18b(example_data_18())
}
