% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day19.R
\name{day19}
\alias{day19}
\alias{solve19a}
\alias{solve19b}
\alias{example_data_19}
\title{Day 19: Aplenty}
\usage{
solve19a(x)

solve19b(x)

example_data_19(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name).}
}
\description{
\href{https://adventofcode.com/2023/day/19}{Aplenty}
}
\examples{
solve19a(example_data_19())
solve19b(example_data_19())
}
