% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day20.R
\name{day20}
\alias{day20}
\alias{f20a}
\alias{f20b}
\alias{example_data_20}
\title{Day 20: Pulse Propagation}
\usage{
f20a(x)

f20b(x)

example_data_20(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f20a(x)} returns .... For Part Two,
\code{f20b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2023/day/20}{Pulse Propagation}
}
\details{
\strong{Part One}

With your help, the Elves manage to find the right parts and fix all of
the machines. Now, they just need to send the command to boot up the
machines and get the sand flowing again.

The machines are far apart and wired together with long \emph{cables}. The
cables don\'t connect to the machines directly, but rather to
communication \emph{modules} attached to the machines that perform various
initialization tasks and also act as communication relays.

Modules communicate using \emph{pulses}. Each pulse is either a \emph{high pulse}
or a \emph{low pulse}. When a module sends a pulse, it sends that type of
pulse to each module in its list of \emph{destination modules}.

There are several different types of modules:

\emph{Flip-flop} modules (prefix \verb{\%}) are either \emph{on} or \emph{off}; they are
initially \emph{off}. If a flip-flop module receives a high pulse, it is
ignored and nothing happens. However, if a flip-flop module receives a
low pulse, it \emph{flips between on and off}. If it was off, it turns on and
sends a high pulse. If it was on, it turns off and sends a low pulse.

\emph{Conjunction} modules (prefix \code{&}) \emph{remember} the type of the most
recent pulse received from \emph{each} of their connected input modules; they
initially default to remembering a \emph{low pulse} for each input. When a
pulse is received, the conjunction module first updates its memory for
that input. Then, if it remembers \emph{high pulses} for all inputs, it sends
a \emph{low pulse}; otherwise, it sends a \emph{high pulse}.

There is a single \emph{broadcast module} (named \code{broadcaster}). When it
receives a pulse, it sends the same pulse to all of its destination
modules.

Here at Desert Machine Headquarters, there is a module with a single
button on it called, aptly, the \emph{button module}. When you push the
button, a single \emph{low pulse} is sent directly to the \code{broadcaster}
module.

After pushing the button, you must wait until all pulses have been
delivered and fully handled before pushing it again. Never push the
button if modules are still processing pulses.

Pulses are always processed \emph{in the order they are sent}. So, if a pulse
is sent to modules \code{a}, \code{b}, and \code{c}, and then module \code{a} processes its
pulse and sends more pulses, the pulses sent to modules \code{b} and \code{c}
would have to be handled first.

The module configuration (your puzzle input) lists each module. The name
of the module is preceded by a symbol identifying its type, if any. The
name is then followed by an arrow and a list of its destination modules.
For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{broadcaster -> a, b, c
\%a -> b
\%b -> c
\%c -> inv
&inv -> a
}\if{html}{\out{</div>}}

In this module configuration, the broadcaster has three destination
modules named \code{a}, \code{b}, and \code{c}. Each of these modules is a flip-flop
module (as indicated by the \verb{\%} prefix). \code{a} outputs to \code{b} which
outputs to \code{c} which outputs to another module named \code{inv}. \code{inv} is a
conjunction module (as indicated by the \code{&} prefix) which, because it
has only one input, acts like an
[inverter]{title="This puzzle originally had a separate inverter module type until I realized it was just a worse conjunction module."}
(it sends the opposite of the pulse type it receives); it outputs to
\code{a}.

By pushing the button once, the following pulses are sent:

\if{html}{\out{<div class="sourceCode">}}\preformatted{button -low-> broadcaster
broadcaster -low-> a
broadcaster -low-> b
broadcaster -low-> c
a -high-> b
b -high-> c
c -high-> inv
inv -low-> a
a -low-> b
b -low-> c
c -low-> inv
inv -high-> a
}\if{html}{\out{</div>}}

After this sequence, the flip-flop modules all end up \emph{off}, so pushing
the button again repeats the same sequence.

Here\'s a more interesting example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{broadcaster -> a
\%a -> inv, con
&inv -> b
\%b -> con
&con -> output
}\if{html}{\out{</div>}}

This module configuration includes the \code{broadcaster}, two flip-flops
(named \code{a} and \code{b}), a single-input conjunction module (\code{inv}), a
multi-input conjunction module (\code{con}), and an untyped module named
\code{output} (for testing purposes). The multi-input conjunction module
\code{con} watches the two flip-flop modules and, if they\'re both on, sends
a \emph{low pulse} to the \code{output} module.

Here\'s what happens if you push the button once:

\if{html}{\out{<div class="sourceCode">}}\preformatted{button -low-> broadcaster
broadcaster -low-> a
a -high-> inv
a -high-> con
inv -low-> b
con -high-> output
b -high-> con
con -low-> output
}\if{html}{\out{</div>}}

Both flip-flops turn on and a low pulse is sent to \code{output}! However,
now that both flip-flops are on and \code{con} remembers a high pulse from
each of its two inputs, pushing the button a second time does something
different:

\if{html}{\out{<div class="sourceCode">}}\preformatted{button -low-> broadcaster
broadcaster -low-> a
a -low-> inv
a -low-> con
inv -high-> b
con -high-> output
}\if{html}{\out{</div>}}

Flip-flop \code{a} turns off! Now, \code{con} remembers a low pulse from module
\code{a}, and so it sends only a high pulse to \code{output}.

Push the button a third time:

\if{html}{\out{<div class="sourceCode">}}\preformatted{button -low-> broadcaster
broadcaster -low-> a
a -high-> inv
a -high-> con
inv -low-> b
con -low-> output
b -low-> con
con -high-> output
}\if{html}{\out{</div>}}

This time, flip-flop \code{a} turns on, then flip-flop \code{b} turns off.
However, before \code{b} can turn off, the pulse sent to \code{con} is handled
first, so it \emph{briefly remembers all high pulses} for its inputs and
sends a low pulse to \code{output}. After that, flip-flop \code{b} turns off,
which causes \code{con} to update its state and send a high pulse to
\code{output}.

Finally, with \code{a} on and \code{b} off, push the button a fourth time:

\if{html}{\out{<div class="sourceCode">}}\preformatted{button -low-> broadcaster
broadcaster -low-> a
a -low-> inv
a -low-> con
inv -high-> b
con -high-> output
}\if{html}{\out{</div>}}

This completes the cycle: \code{a} turns off, causing \code{con} to remember only
low pulses and restoring all modules to their original states.

To get the cables warmed up, the Elves have pushed the button \code{1000}
times. How many pulses got sent as a result (including the pulses sent
by the button itself)?

In the first example, the same thing happens every time the button is
pushed: \code{8} low pulses and \code{4} high pulses are sent. So, after pushing
the button \code{1000} times, \code{8000} low pulses and \code{4000} high pulses are
sent. Multiplying these together gives \emph{\code{32000000}}.

In the second example, after pushing the button \code{1000} times, \code{4250} low
pulses and \code{2750} high pulses are sent. Multiplying these together gives
\emph{\code{11687500}}.

Consult your module configuration; determine the number of low pulses
and high pulses that would be sent after pushing the button \code{1000}
times, waiting for all pulses to be fully handled after each push of the
button. \emph{What do you get if you multiply the total number of low pulses
sent by the total number of high pulses sent?}

\strong{Part Two}

The final machine responsible for moving the sand down to Island Island
has a module attached named \code{rx}. The machine turns on when a \emph{single
low pulse} is sent to \code{rx}.

Reset all modules to their default states. Waiting for all pulses to be
fully handled after each button press, \emph{what is the fewest number of
button presses required to deliver a single low pulse to the module
named \code{rx}?}
}
\examples{
f20a(example_data_20())
f20b()
}
