% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day23.R
\name{day23}
\alias{day23}
\alias{f23a}
\alias{f23b}
\alias{example_data_23}
\title{Day 23: A Long Walk}
\usage{
f23a(x)

f23b(x)

example_data_23(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f23a(x)} returns .... For Part Two,
\code{f23b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2023/day/23}{A Long Walk}
}
\details{
\strong{Part One}

The Elves resume water filtering operations! Clean water starts flowing
over the edge of Island Island.

They offer to help \emph{you} go over the edge of Island Island, too! Just
[hold on tight]{title="It'll be fiiiiiiiine."} to one end of this
impossibly long rope and they\'ll lower you down a safe distance from
the massive waterfall you just created.

As you finally reach Snow Island, you see that the water isn\'t really
reaching the ground: it\'s being \emph{absorbed by the air} itself. It looks
like you\'ll finally have a little downtime while the moisture builds up
to snow-producing levels. Snow Island is pretty scenic, even without any
snow; why not take a walk?

There\'s a map of nearby hiking trails (your puzzle input) that
indicates \emph{paths} (\code{.}), \emph{forest} (\verb{#}), and steep \emph{slopes} (\code{^}, \code{>},
\code{v}, and \code{<}).

For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#.#####################
#.......#########...###
#######.#########.#.###
###.....#.>.>.###.#.###
###v#####.#v#.###.#.###
###.>...#.#.#.....#...#
###v###.#.#.#########.#
###...#.#.#.......#...#
#####.#.#.#######.#.###
#.....#.#.#.......#...#
#.#####.#.#.#########v#
#.#...#...#...###...>.#
#.#.#v#######v###.###v#
#...#.>.#...>.>.#.###.#
#####v#.#.###v#.#.###.#
#.....#...#...#.#.#...#
#.#########.###.#.#.###
#...###...#...#...#.###
###.###.#.###v#####v###
#...#...#.#.>.>.#.>.###
#.###.###.#.###.#.#v###
#.....###...###...#...#
#####################.#
}\if{html}{\out{</div>}}

You\'re currently on the single path tile in the top row; your goal is
to reach the single path tile in the bottom row. Because of all the mist
from the waterfall, the slopes are probably quite \emph{icy}; if you step
onto a slope tile, your next step must be \emph{downhill} (in the direction
the arrow is pointing). To make sure you have the most scenic hike
possible, \emph{never step onto the same tile twice}. What is the longest
hike you can take?

In the example above, the longest hike you can take is marked with \code{O},
and your starting position is marked \code{S}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#S#####################
#OOOOOOO#########...###
#######O#########.#.###
###OOOOO#OOO>.###.#.###
###O#####O#O#.###.#.###
###OOOOO#O#O#.....#...#
###v###O#O#O#########.#
###...#O#O#OOOOOOO#...#
#####.#O#O#######O#.###
#.....#O#O#OOOOOOO#...#
#.#####O#O#O#########v#
#.#...#OOO#OOO###OOOOO#
#.#.#v#######O###O###O#
#...#.>.#...>OOO#O###O#
#####v#.#.###v#O#O###O#
#.....#...#...#O#O#OOO#
#.#########.###O#O#O###
#...###...#...#OOO#O###
###.###.#.###v#####O###
#...#...#.#.>.>.#.>O###
#.###.###.#.###.#.#O###
#.....###...###...#OOO#
#####################O#
}\if{html}{\out{</div>}}

This hike contains \emph{\code{94}} steps. (The other possible hikes you could
have taken were \code{90}, \code{86}, \code{82}, \code{82}, and \code{74} steps long.)

Find the longest hike you can take through the hiking trails listed on
your map. \emph{How many steps long is the longest hike?}

\strong{Part Two}

As you reach the trailhead, you realize that the ground isn\'t as
slippery as you expected; you\'ll have \emph{no problem} climbing up the
steep slopes.

Now, treat all \emph{slopes} as if they were normal \emph{paths} (\code{.}). You still
want to make sure you have the most scenic hike possible, so continue to
ensure that you \emph{never step onto the same tile twice}. What is the
longest hike you can take?

In the example above, this increases the longest hike to \emph{\code{154}} steps:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#S#####################
#OOOOOOO#########OOO###
#######O#########O#O###
###OOOOO#.>OOO###O#O###
###O#####.#O#O###O#O###
###O>...#.#O#OOOOO#OOO#
###O###.#.#O#########O#
###OOO#.#.#OOOOOOO#OOO#
#####O#.#.#######O#O###
#OOOOO#.#.#OOOOOOO#OOO#
#O#####.#.#O#########O#
#O#OOO#...#OOO###...>O#
#O#O#O#######O###.###O#
#OOO#O>.#...>O>.#.###O#
#####O#.#.###O#.#.###O#
#OOOOO#...#OOO#.#.#OOO#
#O#########O###.#.#O###
#OOO###OOO#OOO#...#O###
###O###O#O###O#####O###
#OOO#OOO#O#OOO>.#.>O###
#O###O###O#O###.#.#O###
#OOOOO###OOO###...#OOO#
#####################O#
}\if{html}{\out{</div>}}

Find the longest hike you can take through the surprisingly dry hiking
trails listed on your map. \emph{How many steps long is the longest hike?}
}
\examples{
f23a(example_data_23())
f23b()
}
