% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day25.R
\name{day25}
\alias{day25}
\alias{solve25a}
\alias{example_data_25}
\title{Day 25: Full of Hot Air}
\usage{
solve25a(x)

example_data_25(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name)}
}
\description{
\href{https://adventofcode.com/2022/day/25}{Full of Hot Air}
}
\examples{
solve25a(example_data_25())
}
